Split core.rs into a core/ module
3879 lines in one file became ten, each a logical chunk:

  text.rs         line-break normalization, latinization, line splitting,
                  symbol predicates
  file.rs         parsing and filtering RPG Maker files on disk
  base.rs         the Base struct and its shared state
  translation.rs  parsing translation files and emitting them back
  list.rs         event command list processing
  map.rs / other.rs / system.rs / script.rs / plugin.rs
                  the per-file-kind wrappers

Code moved verbatim; the only edits are import lists, and visibility on the
Base internals that an impl in a sibling module can no longer reach, which
went from private to pub(super)/pub(crate). The thread-local regexes moved
next to their only consumers.

`string_is_only_symbols` and `ends_with_if_index` were associated functions on
Base that touched no state; they are free functions in text.rs now.

Public API is unchanged - core/mod.rs re-exports everything core.rs exported.
