--- src/core.rs
@@ -1,3879 +1,0 @@
-#[allow(unused_imports)]
-use crate::types::ReadMode;
-use crate::{
-    BaseFlags, CommentPos, Comments, IndexSetExt, ProcessedData,
-    constants::{
-        AT_POSITION_MSG, COMMENT_PREFIX, COMMENT_SUFFIX,
-        COULD_NOT_SPLIT_LINE_MSG, EVENT_ID_COMMENT, EVENT_NAME_COMMENT,
-        EVENT_POS_COMMENT, ID_COMMENT, IGNORE_ENTRY_COMMENT, IN_FILE_MSG,
-        INSTANCE_VAR_PREFIX, MAP_DISPLAY_NAME_COMMENT_PREFIX,
-        MAP_ORDER_COMMENT, NAME_COMMENT, NEW_LINE, SEPARATOR, SYMBOLS,
-    },
-    types::{
-        Code, DuplicateMode, EngineType, Error, GameType,
-        IgnoreEntry, IgnoreMap, IndexMapExt, IndexMapGx, Labels, Lines, Mode,
-        RPGMFileType, Scripts, TranslationEntry, TranslationMap, Variable,
-    },
-};
-use flate2::{Compression, read::ZlibDecoder, write::ZlibEncoder};
-use gxhash::{GxBuildHasher, HashMap, HashMapExt, HashSet, HashSetExt};
-use indexmap::map::Entry;
-use log::warn;
-use marshal_rs::{Get, Value, ValueType, dump, load_binary, load_utf8};
-use regex::Regex;
-use serde_json::{Value as SerdeValue, from_str, to_vec};
-use smallvec::{SmallVec, smallvec};
-use std::{
-    borrow::Cow,
-    cell::LazyCell,
-    fmt::Write as FmtWrite,
-    fs::DirEntry,
-    io::{Read, Write},
-    mem::{replace, take, transmute},
-    ops::{ControlFlow, Range},
-    path::Path,
-};
-
-const DISPLAY_NAME_POS: usize = CommentPos::DisplayName as usize;
-
-const BOM: &[u8] = &[0xEF, 0xBB, 0xBF];
-
-/// Newer RPG Maker versions store events in arrays while older versions use hash maps.
-#[repr(u8)]
-enum EventIterator<'a> {
-    New(std::iter::Skip<std::slice::IterMut<'a, Value>>),
-    Old(indexmap::map::ValuesMut<'a, Value, Value>),
-}
-
-impl<'a> Iterator for EventIterator<'a> {
-    type Item = &'a mut Value;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            EventIterator::New(iter) => iter.next(),
-            EventIterator::Old(iter) => iter.next(),
-        }
-    }
-}
-
-thread_local! {
-    static IS_INVALID_MULTILINE_VARIABLE_RE: LazyCell<Regex> =
-        LazyCell::new(|| unsafe {
-            Regex::new(r"^#? ?<.*>.?$|^[a-z]\d$").unwrap_unchecked()
-        });
-    static IS_INVALID_VARIABLE_RE: LazyCell<Regex> = LazyCell::new(|| unsafe {
-        Regex::new(r"^[+-]?$|^///|---|restrict eval").unwrap_unchecked()
-    });
-    static PLUGINS_REGEXPS: LazyCell<[Regex; 11]> = LazyCell::new(|| unsafe {
-        [
-            Regex::new(r"^(name|description|Window Width|Window Height|ATTENTION!!!|Shown Elements|Width|Outline Color|Command Alignment|Command Position|Command Rows|Chinese Font|Korean Font|Default Font|Text Align|Scenes To Draw|displacementImage|Turn Alignment|Buff Formula|Counter Alignment|Default Width|Face Indent|Fast Forward Key|Font Name|Font Name CH|Font Name KR|Name Box Padding|Name Box Added Text|Critical Rate Formula|Critical Multplier Formula|Flat Critical Formula|Default SE|---List---|Button Events List|Kill Switch|Ex Turn Image|Ex Turn Name Color|Non Ex Turn Name Color|Option menu entry|Add to options|Default Ambient Light|Reset Lights|Gab Font Name|Escape Ratio|Translated Format|Default Sound|Action Speed|Default System|Untranslated Format|Default Format|Victory Screen Level Sound|Warning Side Battle UI|Weapon Swap Text Hit|Weapon Swap Text Critical|Weapon Swap Command|Weapon Swap Text Evasion|alwaysDash|renderingMode|Attributes Command|Attributes Column 1|Attributes Column 2|Attributes Column 3|Warning OTB|</span> Minimum Damage</span></td>|Present Settings)$").unwrap_unchecked(),
-            Regex::new(r"^Folder.*\w$").unwrap_unchecked(),
-            Regex::new(r"[XY]$").unwrap_unchecked(),
-            Regex::new(r"BGM").unwrap_unchecked(),
-            Regex::new(r"Label").unwrap_unchecked(),
-            Regex::new(r"^Custom \w").unwrap_unchecked(),
-            Regex::new(r"^outlineColor").unwrap_unchecked(),
-            Regex::new(r"^(Menu|Item|Skill|Equip|Status|Save|Options|End).*(Background|Motion)$").unwrap_unchecked(),
-            Regex::new(r"^Menu \w").unwrap_unchecked(),
-            Regex::new(r"^(MHP|MMP|ATK|DEF|MAT|MDF|AGI|LUK).*(Formula|Maximum|Minimum|Effect|Color)$").unwrap_unchecked(),
-            Regex::new(r"^Damage\w*$").unwrap_unchecked(),
-        ]
-    });
-    static IS_ONLY_SYMBOLS_RE: LazyCell<Regex> = LazyCell::new(|| unsafe {
-        Regex::new(r#"^[,.()+\-:;\[\]^~%&!№$@`*\/→×？?ｘ％▼|♥♪！：〜『』「」〽。…‥＝゠、，【】［］｛｝（）〔〕｟｠〘〙〈〉《》・\\#<>=_ー※▶ⅠⅰⅡⅱⅢⅲⅣⅳⅤⅴⅥⅵⅦⅶⅧⅷⅨⅸⅩⅹⅪⅺⅫⅻⅬⅼⅭⅽⅮⅾⅯⅿ\s\d"']+$"#).unwrap_unchecked()
-    });
-}
-
-pub(crate) trait CustomReplace {
-    /// Normalizes RPG Maker line break symbols (`\n`, `\r`, `\r\n`) to the format that the library uses (`\#`).
-    fn normalize(&self) -> Cow<'_, str>;
-
-    /// Denormalizes library line break symbols to the format that RPG Maker uses (`\n`).
-    fn denormalize(&self) -> Cow<'_, str>;
-}
-
-impl CustomReplace for str {
-    fn normalize(&self) -> Cow<'_, str> {
-        let bytes = self.as_bytes();
-
-        let Some(first) =
-            bytes.iter().position(|&b| b == b'\r' || b == b'\n')
-        else {
-            return Cow::Borrowed(self);
-        };
-
-        // `\r` and `\n` are replaced independently, so a CRLF pair yields two
-        // separators - matching the leftmost-first `\r|\n|\r\n` regex this replaced.
-        let mut out = Vec::with_capacity(self.len() + 8);
-        out.extend_from_slice(&bytes[..first]);
-        out.extend_from_slice(NEW_LINE.as_bytes());
-
-        let mut chunk_start = first + 1;
-
-        for i in chunk_start..bytes.len() {
-            if bytes[i] == b'\r' || bytes[i] == b'\n' {
-                out.extend_from_slice(&bytes[chunk_start..i]);
-                out.extend_from_slice(NEW_LINE.as_bytes());
-                chunk_start = i + 1;
-            }
-        }
-
-        out.extend_from_slice(&bytes[chunk_start..]);
-
-        // SAFETY: `self` is valid UTF-8, and the only bytes rewritten are ASCII
-        // `\r`/`\n`, which cannot occur inside a multi-byte sequence. The
-        // replacement is ASCII too, so the result is still valid UTF-8.
-        Cow::Owned(unsafe { String::from_utf8_unchecked(out) })
-    }
-
-    fn denormalize(&self) -> Cow<'_, str> {
-        // The first `find` doubles as the no-match check, so a string without
-        // separators is scanned once and never copied.
-        let Some(first) = self.find(NEW_LINE) else {
-            return Cow::Borrowed(self);
-        };
-
-        let mut out = String::with_capacity(self.len());
-        out.push_str(&self[..first]);
-        out.push('\n');
-
-        let mut rest = &self[first + NEW_LINE.len()..];
-
-        while let Some(i) = rest.find(NEW_LINE) {
-            out.push_str(&rest[..i]);
-            out.push('\n');
-            rest = &rest[i + NEW_LINE.len()..];
-        }
-
-        out.push_str(rest);
-        Cow::Owned(out)
-    }
-}
-
-/// Parses RPG Maker file from passed content.
-///
-/// # Parameters
-///
-/// - `content` - Content of file to parse.
-/// - `engine_type` - Engine type of the file.
-/// - `file_type` - Type of the file.
-///
-/// # Returns
-///
-/// - [`Value`] - if file was parsed successfully.
-/// - [`Error`] - if unable to deserialize the file.
-///
-/// # Errors
-///
-/// - [`Error::MarshalLoad`] - if unable to load the Marshal data.
-/// - [`Error::JsonParse`] - if unable to parse the JSON data.
-///
-pub fn parse_rpgm_file(
-    mut content: &[u8],
-    engine_type: EngineType,
-    file_type: RPGMFileType,
-) -> Result<Value, Error> {
-    if engine_type.is_new() {
-        // MZ includes Byte Order Mark in files.
-        if content.starts_with(BOM) {
-            content = &content[3..];
-        }
-
-        // SAFETY: JSON is always valid UTF-8.
-        let parsed = from_str::<SerdeValue>(unsafe {
-            std::str::from_utf8_unchecked(content)
-        })?;
-
-        Ok(Value::from(parsed))
-    } else {
-        let loaded = if file_type.is_scripts() {
-            load_binary(content, INSTANCE_VAR_PREFIX)
-        } else {
-            load_utf8(content, INSTANCE_VAR_PREFIX)
-        }?;
-
-        Ok(loaded)
-    }
-}
-
-/// Filters entries of [`std::fs::ReadDir`] and returns iterator of only `Map` entries.
-///
-/// # Parameters
-///
-/// - `entries` - Entries read with [`std::fs::read_dir`].
-/// - `engine_extension` - [`&str`] corresponding to the extension of read entries.
-///
-/// # Returns
-///
-/// Filtered iterator containing only `Map` entries.
-///
-pub fn filter_maps<'a>(
-    entries: impl Iterator<Item = &'a DirEntry>,
-    engine_extension: &'a str,
-) -> impl Iterator<Item = &'a DirEntry> {
-    let mut result: Vec<&'a DirEntry> = entries
-        .filter_map(move |entry| {
-            if !entry.file_type().ok()?.is_file() {
-                return None;
-            }
-
-            let filename = entry.file_name();
-            let extension = Path::new(&filename).extension()?;
-            let filename_str = filename.to_str()?;
-
-            if filename_str.starts_with("Map")
-                && filename_str.as_bytes().get(3)?.is_ascii_digit()
-                && extension == engine_extension
-            {
-                return Some(entry);
-            }
-
-            None
-        })
-        .collect();
-
-    result.sort_by_key(|entry| {
-        let filename = entry.file_name();
-        let filename_str = filename.to_str().unwrap_or("");
-        let digits: String = filename_str[3..]
-            .chars()
-            .take_while(|c| c.is_ascii_digit())
-            .collect();
-        digits.parse::<u32>().unwrap_or(0)
-    });
-
-    result.into_iter()
-}
-
-/// Filters entries of [`std::fs::ReadDir`] and returns iterator of only other entries.
-///
-/// # Parameters
-///
-/// - `entries` - Entries read with [`std::fs::read_dir`].
-/// - `engine_extension` - [`&str`] corresponding to the extension of read entries.
-/// - `game_type` - [`GameType`] of entries.
-///
-/// # Returns
-///
-/// Filtered iterator containing only other entries.
-///
-pub fn filter_other<'a>(
-    entries: impl Iterator<Item = &'a DirEntry>,
-    engine_extension: &'a str,
-    game_type: GameType,
-) -> impl Iterator<Item = &'a DirEntry> {
-    let mut result: Vec<&'a DirEntry> = entries
-        .filter_map(move |entry| {
-            if !entry.file_type().ok()?.is_file() {
-                return None;
-            }
-            let filename = entry.file_name();
-            let filename_path = Path::new(&filename);
-            let basename = filename_path
-                .file_stem()
-                .and_then(|basename| basename.to_str())?;
-            let extension = filename_path.extension()?;
-            let file_type = RPGMFileType::from_filename(basename);
-            if extension == engine_extension && file_type.is_other() {
-                if game_type.is_termina() && file_type.is_states() {
-                    return None;
-                }
-                return Some(entry);
-            }
-            None
-        })
-        .collect();
-
-    result.sort_by_key(|entry| entry.file_name());
-    result.into_iter()
-}
-
-/// Parses ignore file contents to [`IgnoreMap`].
-///
-/// # Parameters
-///
-/// - `ignore_file_path` - Path to the `.rvpacker-ignore` file.
-/// - `duplicate_mode` - [`DuplicateMode`], which was used during read.
-/// - `read` - Parse for reading or purging.
-///
-/// # Returns
-///
-/// Parsed [`IgnoreMap`].
-///
-#[must_use]
-pub fn parse_ignore(
-    ignore_file_content: &str,
-    duplicate_mode: DuplicateMode,
-    read: bool,
-) -> IgnoreMap {
-    let mut ignore_map = IgnoreMap::default();
-    let mut ignore_file_lines = ignore_file_content.lines();
-
-    let Some(mut first_entry_comment) = ignore_file_lines.next() else {
-        return ignore_map;
-    };
-
-    if read
-        && duplicate_mode.is_remove()
-        && !(first_entry_comment.contains("<#>System")
-            || first_entry_comment.contains("<#>Scripts")
-            || first_entry_comment.contains("<#>Plugins"))
-    {
-        // If duplicates are removed, we should group all ignore entries
-        // that correspond to a single file into one ignore entry.
-        first_entry_comment = &first_entry_comment
-            [..unsafe { first_entry_comment.find(':').unwrap_unchecked() }];
-    }
-
-    ignore_map.reserve_exact(256);
-    ignore_map.insert(
-        first_entry_comment.to_string(),
-        IgnoreEntry::with_capacity(128),
-    );
-
-    let mut ignore_entry =
-        unsafe { ignore_map.last_mut().unwrap_unchecked().1 };
-
-    for mut line in ignore_file_lines.filter(|line| !line.is_empty()) {
-        if let Some(mid) = line.strip_prefix(IGNORE_ENTRY_COMMENT) {
-            // If duplicates are allowed, we should group all ignore entries
-            // that correspond to a single file into one ignore entry.
-            if read
-                && duplicate_mode.is_remove()
-                && !(mid.starts_with("<#>System")
-                    || mid.starts_with("<#>Scripts")
-                    || mid.starts_with("<#>Plugins"))
-            {
-                line = &mid[..unsafe { mid.find(':').unwrap_unchecked() }];
-            }
-
-            ignore_map
-                .entry(line.into())
-                .or_insert(IgnoreEntry::with_capacity(128));
-            ignore_entry =
-                unsafe { ignore_map.last_mut().unwrap_unchecked().1 };
-        } else {
-            ignore_entry.insert(line.into());
-        }
-    }
-
-    ignore_map
-}
-
-/// Extracts the game title from a `Game.ini` file's content.
-///
-/// # Parameters
-///
-/// - `ini_file_content` - raw byte content of the INI file to parse.
-///
-/// # Returns
-///
-/// - [`Vec<u8>`] - vector of extracted title's bytes on success. Title may not be UTF-8.
-/// - [`Error`] - otherwise.
-///
-/// # Errors
-///
-/// - [`Error::NoTitle`] - if no "Title" entry is found in the INI file.
-///
-/// # Example
-///
-/// ```no_run
-/// use rvpacker_txt_rs_lib::{get_ini_title, Error};
-/// use std::fs::read;
-///
-/// fn main() -> Result<(), Box<dyn std::error::Error>> {
-///     let ini_content = read("C:/Game/Game.ini")?;
-///     let title = get_ini_title(&ini_content)?;
-///     Ok(())
-/// }
-/// ```
-pub fn get_ini_title(ini_file_content: &[u8]) -> Result<Vec<u8>, Error> {
-    fn trim_bytes(bytes: &[u8]) -> &[u8] {
-        let start = bytes.iter().position(|&b| !is_space(b)).unwrap_or(0);
-        let end = bytes
-            .iter()
-            .rposition(|&b| !is_space(b))
-            .map_or(0, |i| i + 1);
-        &bytes[start..end]
-    }
-
-    fn is_space(b: u8) -> bool {
-        b == b' ' || b == b'\t' || b == b'\r'
-    }
-
-    fn split_lines(data: &[u8]) -> SmallVec<[&[u8]; 4]> {
-        let mut lines = SmallVec::with_capacity(4);
-        let mut start = 0;
-        let mut i = 0;
-
-        while i < data.len() {
-            if data[i] == b'\n' {
-                lines.push(&data[start..i]);
-                i += 1;
-                start = i;
-            } else if data[i] == b'\r' {
-                lines.push(&data[start..i]);
-
-                if data.get(i + 1).is_some_and(|ch| *ch == b'\n') {
-                    i += 2;
-                } else {
-                    i += 1;
-                }
-
-                start = i;
-            } else {
-                i += 1;
-            }
-        }
-
-        if start < data.len() {
-            lines.push(&data[start..]);
-        }
-
-        lines
-    }
-
-    for line in split_lines(ini_file_content) {
-        if line.to_ascii_lowercase().starts_with(b"title") {
-            if let Some(pos) = line.iter().position(|&b| b == b'=') {
-                let right = &line[pos + 1..];
-                let trimmed = trim_bytes(right);
-                return Ok(trimmed.to_vec());
-            }
-        }
-    }
-
-    Err(Error::NoTitle)
-}
-
-/// Extracts the game title from a `System.json` file's content.
-///
-/// # Parameters
-///
-/// - `system_file_content` - JSON string content of the system file
-///
-/// # Returns
-///
-/// - [`String`] game title extracted from the "gameTitle" field if successful.
-/// - [`Error`] otherwise.
-///
-/// # Errors
-///
-/// - [`Error::JsonParse`] - if parsing `system_file_content` failed.
-/// - [`Error::NoTitle`] - if the parsed JSON doesn't contain "gameTitle" key.
-///
-/// # Example
-///
-/// ```no_run
-/// use rvpacker_txt_rs_lib::{get_system_title, Error};
-/// use std::fs::read_to_string;
-///
-/// fn main() -> Result<(), Box<dyn std::error::Error>> {
-///     let system_file_content = read_to_string("C:/Game/www/data/System.json")?;
-///     let title = get_system_title(&system_file_content)?;
-///     Ok(())
-/// }
-/// ```
-pub fn get_system_title(
-    mut system_file_content: &str,
-) -> Result<String, Error> {
-    // MZ includes Byte Order Mark in files.
-    if system_file_content.as_bytes().starts_with(BOM) {
-        system_file_content = &system_file_content[BOM.len()..];
-    }
-
-    let system_file_value: SerdeValue = from_str(system_file_content)?;
-
-    system_file_value["gameTitle"]
-        .as_str()
-        .map(Into::into)
-        .ok_or(Error::NoTitle)
-}
-
-/// Latinize CJK/Unicode (mainly punctuation) characters to their Latin equivalents.
-///
-/// # Parameters
-///
-/// - `string` - String to latinize.
-///
-/// # Returns
-///
-/// - [`Cow<str>`] - as owned if replacements occurred, as borrowed otherwise.
-///
-pub fn latinize_string(string: &str) -> Cow<'_, str> {
-    let mut result: Option<String> = None;
-
-    for (i, char) in string.chars().enumerate() {
-        let replacement = match char {
-            '。' => ".",
-            '、' | '，' => ",",
-            '・' | '※' => "·",
-            '゠' => "–",
-            '＝' | 'ー' => "—",
-            '「' | '」' | '〈' | '〉' => "'",
-            '『' | '』' | '《' | '》' => "\"",
-            '（' | '〔' | '｟' | '〘' => "(",
-            '）' | '〕' | '｠' | '〙' => ")",
-            '｛' => "{",
-            '｝' => "}",
-            '［' | '【' | '〖' | '〚' => "[",
-            '］' | '】' | '〗' | '〛' => "]",
-            '〜' => "~",
-            '？' => "?",
-            '！' => "!",
-            '：' => ":",
-            '…' | '‥' => "...",
-            '　' => " ",
-            'Ⅰ' => "I",
-            'ⅰ' => "i",
-            'Ⅱ' => "II",
-            'ⅱ' => "ii",
-            'Ⅲ' => "III",
-            'ⅲ' => "iii",
-            'Ⅳ' => "IV",
-            'ⅳ' => "iv",
-            'Ⅴ' => "V",
-            'ⅴ' => "v",
-            'Ⅵ' => "VI",
-            'ⅵ' => "vi",
-            'Ⅶ' => "VII",
-            'ⅶ' => "vii",
-            'Ⅷ' => "VIII",
-            'ⅷ' => "viii",
-            'Ⅸ' => "IX",
-            'ⅸ' => "ix",
-            'Ⅹ' => "X",
-            'ⅹ' => "x",
-            'Ⅺ' => "XI",
-            'ⅺ' => "xi",
-            'Ⅻ' => "XII",
-            'ⅻ' => "xii",
-            'Ⅼ' => "L",
-            'ⅼ' => "l",
-            'Ⅽ' => "C",
-            'ⅽ' => "c",
-            'Ⅾ' => "D",
-            'ⅾ' => "d",
-            'Ⅿ' => "M",
-            'ⅿ' => "m",
-            _ => {
-                if let Some(s) = &mut result {
-                    s.push(char);
-                }
-                continue;
-            }
-        };
-
-        if result.is_none() {
-            let mut s = String::with_capacity(string.len());
-            s.push_str(&string[..string.char_indices().nth(i).unwrap().0]);
-            result = Some(s);
-        }
-
-        result.as_mut().unwrap().push_str(replacement);
-    }
-
-    match result {
-        Some(s) => Cow::Owned(s),
-        None => Cow::Borrowed(string),
-    }
-}
-
-/// Outcome of splitting one `source<#>translation` line from a translation file.
-pub(crate) enum TranslationLine<'a> {
-    /// No separator present; the caller should warn and skip the line.
-    Malformed,
-    /// Empty translation on a write pass - the entry is unused, so skip it.
-    Untranslated,
-    Split {
-        source: Cow<'a, str>,
-        translation: Cow<'a, str>,
-    },
-}
-
-/// Splits one translation-file line into its source and translation halves.
-///
-/// The translation is the last non-empty field after the first separator, so files
-/// carrying several translation columns still resolve to the rightmost filled one.
-///
-/// Borrows throughout; allocates only where `write` forces denormalization.
-pub(crate) fn split_translation_line(
-    line: &str,
-    trim: bool,
-    write: bool,
-) -> TranslationLine<'_> {
-    let Some((source, rest)) = line.split_once(SEPARATOR) else {
-        return TranslationLine::Malformed;
-    };
-
-    let translation = rest
-        .rsplit(SEPARATOR)
-        .find(|field| !field.is_empty())
-        .unwrap_or_default();
-
-    let (source, translation) = if trim {
-        (source.trim(), translation.trim())
-    } else {
-        (source, translation)
-    };
-
-    if write {
-        // Lines with no translation are unused on write.
-        if translation.is_empty() {
-            return TranslationLine::Untranslated;
-        }
-
-        TranslationLine::Split {
-            source: source.denormalize(),
-            translation: translation.denormalize(),
-        }
-    } else {
-        TranslationLine::Split {
-            source: Cow::Borrowed(source),
-            translation: Cow::Borrowed(translation),
-        }
-    }
-}
-
-pub(crate) fn push_metadata(
-    output: &mut Vec<u8>,
-    id: u16,
-    metadata: &[String],
-) {
-    output.extend_from_slice(ID_COMMENT.as_bytes());
-    output.extend_from_slice(SEPARATOR.as_bytes());
-    output.extend_from_slice(id.to_string().as_bytes());
-    output.push(b'\n');
-
-    for comment in metadata.iter().filter(|c| !c.is_empty()) {
-        output.extend_from_slice(comment.as_bytes());
-        output.push(b'\n');
-    }
-}
-
-pub(crate) fn push_entries(
-    output: &mut Vec<u8>,
-    source: &str,
-    translation: &TranslationEntry,
-) {
-    for comment in translation.comments.iter().filter(|c| !c.is_empty()) {
-        output.extend_from_slice(comment.as_bytes());
-        output.push(b'\n');
-    }
-
-    if !source.is_empty() {
-        output.extend_from_slice(source.as_bytes());
-        output.extend_from_slice(SEPARATOR.as_bytes());
-    }
-
-    if !translation.is_empty() {
-        output.extend_from_slice(translation.as_bytes());
-    }
-
-    if !source.is_empty() || !translation.is_empty() {
-        output.push(b'\n');
-    }
-}
-
-/// Lines flushed for one accumulated entry.
-///
-/// The duplicate-removing path leaves its lines in [`Base::lines`] and records only
-/// the range they occupy, so they are never copied; the duplicate-allowing path
-/// drains them out and owns them.
-enum FlushedLines {
-    Owned(Vec<String>),
-    Range(Range<usize>),
-}
-
-impl FlushedLines {
-    const EMPTY: Self = Self::Range(0..0);
-
-    fn len(&self) -> usize {
-        match self {
-            Self::Owned(lines) => lines.len(),
-            Self::Range(range) => range.len(),
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// `store` must be the [`Base::lines`] these were flushed from.
-    fn get<'a>(&'a self, store: &'a Lines, index: usize) -> &'a str {
-        match self {
-            Self::Owned(lines) => &lines[index],
-            Self::Range(range) => &store[range.start + index],
-        }
-    }
-}
-
-pub struct Base {
-    pub mode: Mode,
-    pub flags: BaseFlags,
-    pub game_type: GameType,
-    pub engine_type: EngineType,
-    pub duplicate_mode: DuplicateMode,
-
-    pub skip_maps: HashSet<u16>,
-    pub skip_events: HashMap<RPGMFileType, HashSet<u16>>,
-    pub map_events: bool,
-
-    pub ignore_map: IgnoreMap,
-
-    /// Index of the currently selected entry in `ignore_map`, or `usize::MAX` when
-    /// none is selected. Only set when the ignore flags are on; an index rather than
-    /// a reference because the entry is owned by this same struct.
-    ignore_entry_index: usize,
-
-    translation_initialized: bool,
-
-    lines: Lines,
-    total_length: usize,
-
-    metadata: HashMap<u16, Comments>,
-
-    translation_maps: IndexMapGx<u16, TranslationMap>,
-
-    /// Index of the currently selected map in `translation_maps`, or `usize::MAX`
-    /// when none is selected. An index rather than a reference because the map it
-    /// points at is owned by this same struct; `translation_maps` is only ever
-    /// appended to while a selection is live, so indices stay valid.
-    translation_map_index: usize,
-
-    /// Flattened `translation_maps`, built once on write with [`DuplicateMode::Remove`].
-    /// See [`Base::build_write_lookup`].
-    write_lookup: TranslationMap,
-
-    accumulated_translation:
-        Vec<(u16, Comments, FlushedLines, TranslationMap)>,
-    top_level_comments: HashMap<u16, Vec<String>>,
-
-    file_type: RPGMFileType,
-    labels: Labels,
-}
-
-impl Default for Base {
-    fn default() -> Self {
-        Self {
-            mode: Mode::Read(ReadMode::Default { force: false }),
-            flags: BaseFlags::empty(),
-            game_type: GameType::None,
-            engine_type: EngineType::New,
-            duplicate_mode: DuplicateMode::Remove,
-
-            ignore_map: IgnoreMap::default(),
-            ignore_entry_index: usize::MAX,
-
-            translation_initialized: false,
-
-            lines: Lines::default(),
-            total_length: 0,
-
-            metadata: HashMap::default(),
-
-            translation_map_index: usize::MAX,
-            translation_maps: IndexMapGx::default(),
-            write_lookup: TranslationMap::default(),
-
-            accumulated_translation: Vec::new(),
-            top_level_comments: HashMap::default(),
-
-            map_events: false,
-            file_type: RPGMFileType::Invalid,
-            labels: Labels::default(),
-
-            skip_maps: HashSet::default(),
-            skip_events: HashMap::default(),
-        }
-    }
-}
-
-impl Base {
-    /// Creates new base from mode and engine type.
-    ///
-    /// # Parameters
-    ///
-    /// - `mode` - [`Mode`] to use.
-    /// - `engine_type` - [`EngineType`] to use.
-    ///
-    #[must_use]
-    pub fn new(mode: Mode, engine_type: EngineType) -> Self {
-        Self {
-            mode,
-            engine_type,
-            labels: Labels::new(engine_type),
-            lines: Lines::with_capacity(512),
-
-            metadata: HashMap::with_capacity(1024),
-            translation_maps: IndexMapGx::with_capacity(1024),
-
-            ..Default::default()
-        }
-    }
-
-    /// Clears all the underlying collections, and makes this base ready to be used in the next base.
-    ///
-    /// This function is used by file-specific bases' constructors, so you generally mustn't call it manually.
-    pub fn reset(&mut self) {
-        self.translation_initialized = false;
-
-        self.lines.clear();
-        self.total_length = 0;
-
-        self.metadata.clear();
-
-        self.translation_maps.clear();
-        self.write_lookup.clear();
-        self.accumulated_translation.clear();
-    }
-
-    fn process_parameter(
-        &self,
-        code: Code,
-        mut parameter: &str,
-    ) -> Option<String> {
-        if Self::string_is_only_symbols(parameter) {
-            return None;
-        }
-
-        let mut extra_strings: SmallVec<[(&str, bool); 4]> =
-            SmallVec::with_capacity(4);
-        let mut shop_prefix: Option<&str> = None;
-
-        match self.game_type {
-            GameType::Termina => {
-                if parameter.chars().all(|c| {
-                    c.is_ascii_lowercase()
-                        || (c.is_ascii_punctuation() && c != '"')
-                }) {
-                    return None;
-                }
-
-                if code.is_system()
-                    && !parameter.starts_with("Gab")
-                    && (!parameter.starts_with("choice_text")
-                        || parameter.ends_with("????"))
-                {
-                    return None;
-                }
-            }
-            GameType::LisaRPG => {
-                if code.is_any_dialogue() {
-                    if let Some(i) = Self::find_lisa_prefix_index(parameter) {
-                        if Self::string_is_only_symbols(&parameter[i..]) {
-                            return None;
-                        }
-
-                        if self.mode.is_write() {
-                            extra_strings.push((&parameter[..i], false));
-                        }
-
-                        if !parameter.starts_with(r"\et") {
-                            parameter = &parameter[i..];
-                        }
-                    }
-                }
-            }
-            // custom processing for other games
-            GameType::None => {}
-        }
-
-        if !self.engine_type.is_new() {
-            if let Some(i) = Self::ends_with_if_index(parameter) {
-                if self.mode.is_write() {
-                    extra_strings.push((&parameter[..i], true));
-                }
-
-                parameter = &parameter[..i];
-            }
-
-            if code.is_shop() {
-                if !parameter.contains("shop_talk") {
-                    return None;
-                }
-
-                // SAFETY: At this point, shop parameter should always contain '='.
-                let (left, mut actual_string) =
-                    unsafe { parameter.split_once('=').unwrap_unchecked() };
-                shop_prefix = Some(left);
-                actual_string = actual_string.trim();
-
-                if actual_string.len() < 2 {
-                    return None;
-                }
-
-                let without_quotes = &actual_string[1..actual_string.len() - 1];
-
-                if without_quotes.is_empty()
-                    || Self::string_is_only_symbols(without_quotes)
-                {
-                    return None;
-                }
-
-                parameter = without_quotes;
-            }
-        }
-
-        if self.mode.is_write() {
-            self.get_key(parameter).map(|t| {
-                let mut translation = if extra_strings.is_empty() {
-                    t.translation.to_string()
-                } else {
-                    String::new()
-                };
-
-                for (string, append) in extra_strings {
-                    if append {
-                        translation = t.to_string() + string;
-                    } else {
-                        translation = format!("{string}{t}", t = t.translation);
-                    }
-                }
-
-                // Put the `shop_talk_xxx=` prefix back on, so that the result is
-                // self-contained and the caller needn't keep the source string
-                // borrowed while it writes to the value it came from.
-                if let Some(prefix) = shop_prefix {
-                    translation = format!("{prefix}=\"{translation}\"");
-                }
-
-                translation
-            })
-        } else {
-            Some(parameter.to_string())
-        }
-    }
-
-    /// Applies an already-processed parameter, produced by [`Base::process_parameter`].
-    fn process_param(&mut self, value: &mut Value, parsed: String) {
-        if self.mode.is_write() {
-            *value =
-                Self::make_string_value(&parsed, self.engine_type.is_new());
-        } else {
-            self.insert_string(Cow::Owned(parsed));
-        }
-    }
-
-    /// Inserts `string` to `self.lines` if `self.mode`.
-    ///
-    /// Will skip inserting if `self.mode` is not [`Mode::Write`] or `self.flags` contain [`BaseFlags::Ignore`] and `self.ignore_entry` contains the string.
-    ///
-    /// # Parameters
-    ///
-    /// - `string` - String to insert in `self.lines`.
-    ///
-    fn insert_string(&mut self, string: Cow<'_, str>) {
-        if self.mode.is_write()
-            || (self.flags.contains(BaseFlags::Ignore)
-                && self
-                    .ignore_entry()
-                    .is_some_and(|entry| entry.contains(string.as_ref())))
-        {
-            return;
-        }
-
-        self.lines.insert(string.into_owned());
-    }
-
-    fn join_dialogue_lines(
-        &mut self,
-        list: &mut [Value],
-        dialogue_lines: &mut SmallVec<[String; 4]>,
-        dialogue_line_indices: &mut SmallVec<[usize; 4]>,
-        write_string_literally: bool,
-    ) {
-        let joined = dialogue_lines.join(if self.mode.is_write() {
-            "\n"
-        } else {
-            NEW_LINE
-        });
-
-        if self.mode.is_write() {
-            let Some(translation) =
-                self.process_parameter(Code::Dialogue, &joined)
-            else {
-                return;
-            };
-
-            let translation_lines: Vec<&str> = translation.lines().collect();
-            let split_line_count = translation_lines.len();
-            let dialogue_line_count = dialogue_lines.len();
-
-            for (i, &index) in dialogue_line_indices.iter().enumerate() {
-                list[index][self.labels.parameters][0] = if i < split_line_count
-                {
-                    Self::make_string_value(
-                        translation_lines[i],
-                        write_string_literally,
-                    )
-                } else {
-                    // Overwrite leftover source text
-                    Value::string(" ")
-                }
-            }
-
-            if split_line_count > dialogue_line_count {
-                let remaining =
-                    translation_lines[dialogue_line_count - 1..].join("\n");
-
-                // SAFETY: We checked that `dialogue_lines` are not empty before calling this.
-                list[unsafe {
-                    *dialogue_line_indices.last().unwrap_unchecked()
-                }][self.labels.parameters][0] = Value::string(remaining);
-            }
-        } else if let Some(parsed) =
-            self.process_parameter(Code::Dialogue, &joined)
-        {
-            self.process_param(&mut Value::default(), parsed);
-        }
-    }
-
-    /// Processes the list of objects found in `Map`, `CommonEvents` and `Troops` files.
-    ///
-    /// # Parameters
-    ///
-    /// - `list` - list of [`Value`]s.
-    ///
-    fn process_list(&mut self, list: &mut Vec<Value>) {
-        let mut in_sequence = false;
-        let mut write_string_literally = self.engine_type.is_new();
-        let mut dialogue_lines = SmallVec::with_capacity(4);
-        let mut dialogue_line_indices = SmallVec::with_capacity(4);
-
-        // Indexed rather than iterated, because `join_dialogue_lines` needs `list`
-        // itself while we're partway through it.
-        for item_idx in 0..list.len() {
-            let item = &mut list[item_idx];
-
-            // SAFETY: Each item must contain code.
-            let code = Code::from(unsafe {
-                item[self.labels.code].as_int().unwrap_unchecked()
-            } as u16);
-
-            let code = if code.is_dialogue_start() && !self.engine_type.is_xp()
-            {
-                Code::Bad
-            } else {
-                code
-            };
-
-            if self.mode.is_write() && !self.engine_type.is_new() {
-                // SAFETY: Each item must contain parameters.
-                let parameters = unsafe {
-                    item[self.labels.parameters].as_array().unwrap_unchecked()
-                };
-
-                if !parameters.is_empty() {
-                    write_string_literally = !match code {
-                        Code::ChoiceArray => parameters[0][0].is_bytes(),
-                        Code::Misc1 | Code::Misc2 | Code::Choice => {
-                            parameters[1].is_bytes()
-                        }
-                        _ => parameters[0].is_bytes(),
-                    }
-                }
-            }
-
-            if in_sequence
-                && (!self.engine_type.is_xp() && !code.is_any_dialogue())
-                || (code.is_dialogue_start() && !dialogue_lines.is_empty())
-            {
-                if !dialogue_lines.is_empty() {
-                    self.join_dialogue_lines(
-                        list,
-                        &mut dialogue_lines,
-                        &mut dialogue_line_indices,
-                        write_string_literally,
-                    );
-                    dialogue_lines.clear();
-                    dialogue_line_indices.clear();
-                }
-
-                in_sequence = false;
-            }
-
-            if code.is_bad() {
-                continue;
-            }
-
-            // SAFETY: Each item must contain parameters.
-            let parameters = unsafe {
-                list[item_idx][self.labels.parameters]
-                    .as_array_mut()
-                    .unwrap_unchecked()
-            };
-
-            if parameters.is_empty() {
-                continue;
-            }
-
-            let value_index =
-                usize::from(code.is_any_misc() || code.is_choice());
-
-            if code.is_choice_array() {
-                // SAFETY: We have just checked - it's an array.
-                let choices = unsafe {
-                    parameters[value_index].as_array_mut().unwrap_unchecked()
-                };
-
-                for choice_idx in 0..choices.len() {
-                    // Scoped so that the borrow of `choices[choice_idx]` ends
-                    // before we write back into it; `process_parameter` returns
-                    // owned data, so nothing outlives the scope.
-                    let parsed = {
-                        let Some(string) =
-                            self.extract_string(&choices[choice_idx], true)
-                        else {
-                            continue;
-                        };
-
-                        self.process_parameter(code, string)
-                    };
-
-                    if let Some(parsed) = parsed {
-                        self.process_param(&mut choices[choice_idx], parsed);
-                    }
-                }
-            } else {
-                let value = &mut parameters[value_index];
-
-                let Some(parameter_string) =
-                    self.extract_string(&*value, false)
-                else {
-                    continue;
-                };
-
-                if !code.is_credit() && parameter_string.is_empty() {
-                    continue;
-                }
-
-                if code.is_any_dialogue() {
-                    dialogue_lines.push(parameter_string.into());
-
-                    if self.mode.is_write() {
-                        dialogue_line_indices.push(item_idx);
-                    }
-
-                    in_sequence = true;
-                } else if let Some(parsed) =
-                    self.process_parameter(code, parameter_string)
-                {
-                    self.process_param(value, parsed);
-                }
-            }
-        }
-    }
-
-    /// Gets ignore entry from `self.ignore_map` by `id`.
-    ///
-    /// Skips getting an entry if `self.flags` do not contain [`BaseFlags::Ignore`] or [`BaseFlags::CreateIgnore`].
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - ID of the entry to get.
-    ///
-    fn get_ignore_entry(&mut self, id: u16) {
-        if !self
-            .flags
-            .intersects(BaseFlags::CreateIgnore | BaseFlags::Ignore)
-        {
-            return;
-        }
-
-        // Built in one pass. This previously formatted the name, truncated it at
-        // the ':', then formatted the whole key again - two allocations per map
-        // and per event.
-        let mut key = String::with_capacity(
-            IGNORE_ENTRY_COMMENT.len() + SEPARATOR.len() + 24,
-        );
-
-        key.push_str(IGNORE_ENTRY_COMMENT);
-        key.push_str(SEPARATOR);
-        let _ = write!(key, "{file}", file = self.file_type);
-
-        // With duplicates removed the whole file shares a single entry, so the id
-        // is left off.
-        if !(self.flags.contains(BaseFlags::Ignore)
-            && self.duplicate_mode.is_remove())
-        {
-            let _ = write!(key, ": {id}");
-        }
-
-        let entry = self.ignore_map.entry(key);
-
-        self.ignore_entry_index = entry.index();
-        entry.or_default();
-    }
-
-    /// The currently selected ignore entry, or [`None`] if the ignore flags are off
-    /// and no entry was ever selected.
-    fn ignore_entry(&self) -> Option<&IgnoreEntry> {
-        self.ignore_map
-            .get_index(self.ignore_entry_index)
-            .map(|(_, entry)| entry)
-    }
-
-
-    /// Initializes translation by filling `self.translation_maps` with parsed maps from `translation`.
-    ///
-    /// # Parameters
-    ///
-    /// - `translation` - translation file content to parse.
-    ///
-    fn initialize_translation(
-        &mut self,
-        translation: Option<&str>,
-    ) -> Result<(), Error> {
-        if self.mode.is_default() || self.translation_initialized {
-            return Ok(());
-        }
-
-        let Some(translation) = translation else {
-            return Err(Error::NoTranslation);
-        };
-
-        self.translation_initialized = true;
-
-        let trim = if self.file_type.is_main() {
-            self.flags.contains(BaseFlags::Trim)
-        } else {
-            false
-        };
-
-        let mut scratch = TranslationMap::default();
-        let mut translation_lines = translation.lines().enumerate();
-
-        if self.game_type.is_termina() && self.file_type.is_items() {
-            for _ in 0..4 {
-                let (_, item_category_line) =
-                    unsafe { translation_lines.next().unwrap_unchecked() };
-
-                if item_category_line.starts_with("<Menu Category") {
-                    let (source, translation) = unsafe {
-                        item_category_line
-                            .split_once(SEPARATOR)
-                            .unwrap_unchecked()
-                    };
-
-                    scratch.insert(source.into(), translation.into());
-                } else {
-                    panic!(
-                        "items.txt in Fear & Hunger 2: Termina should start \
-                         with 4 `Menu Category` entries."
-                    );
-                }
-            }
-
-            self.translation_maps
-                .insert(u16::MAX, scratch.drain(..).collect());
-        }
-
-        let mut top_level_comments: Vec<String> = Vec::new();
-        let mut comments: Comments = smallvec![String::new(); 3];
-        let mut id = 0;
-        let mut first = true;
-
-        for (i, line) in translation_lines {
-            if line.starts_with(ID_COMMENT) {
-                if id != 0 {
-                    if scratch.is_empty() {
-                        let metadata_entry = self.metadata.entry(id).or_insert(
-                            replace(&mut comments, smallvec![String::new(); 3]),
-                        );
-
-                        let display_name = &metadata_entry[DISPLAY_NAME_POS];
-
-                        if self.mode.is_write()
-                            && (display_name.is_empty()
-                                || display_name.ends_with(SEPARATOR))
-                        {
-                            continue;
-                        }
-
-                        self.translation_maps
-                            .entry(id)
-                            .or_insert(TranslationMap::with_capacity(512));
-                    }
-
-                    self.translation_maps
-                        .insert(id, scratch.drain(..).collect());
-                }
-
-                id = line
-                    .strip_prefix(ID_COMMENT)
-                    .and_then(|n| n.strip_prefix(SEPARATOR))
-                    .and_then(|n| n.trim_end().parse::<u16>().ok())
-                    .unwrap();
-                first = true;
-                comments = smallvec![String::new(); 3];
-                top_level_comments = Vec::new();
-
-                continue;
-            }
-
-            if line.starts_with(COMMENT_PREFIX) {
-                if [EVENT_ID_COMMENT, EVENT_NAME_COMMENT, EVENT_POS_COMMENT]
-                    .into_iter()
-                    .any(|c| line.starts_with(c))
-                {
-                    continue;
-                }
-
-                if first {
-                    let pos = CommentPos::from_str(line);
-
-                    if pos == CommentPos::None {
-                        top_level_comments.push(line.to_string());
-                        continue;
-                    }
-
-                    if pos == CommentPos::DisplayName {
-                        let suffix_pos = line.rfind(COMMENT_SUFFIX).unwrap();
-                        let prefix_len = MAP_DISPLAY_NAME_COMMENT_PREFIX.len();
-                        let source = &line[prefix_len..suffix_pos];
-                        let translation =
-                            line.rsplit_once(SEPARATOR).unwrap().1;
-                        comments[pos as usize] =
-                            format!("{source}{SEPARATOR}{translation}");
-                    } else {
-                        comments[pos as usize] =
-                            line.split_once(SEPARATOR).unwrap().1.to_string();
-                    }
-                } else {
-                    comments.push(line.to_string());
-                }
-
-                continue;
-            }
-
-            let (source, translation) =
-                match split_translation_line(line, trim, self.mode.is_write())
-                {
-                    TranslationLine::Split {
-                        source,
-                        translation,
-                    } => (source, translation),
-                    TranslationLine::Untranslated => continue,
-                    TranslationLine::Malformed => {
-                        warn!(
-                            "{COULD_NOT_SPLIT_LINE_MSG}\n{AT_POSITION_MSG}: \
-                             {i}\n{IN_FILE_MSG}: {file}.txt",
-                            i = i + 1,
-                            file = self.file_type.to_string().to_lowercase()
-                        );
-                        comments = smallvec![String::new(); 3];
-                        continue;
-                    }
-                };
-
-            if first {
-                self.top_level_comments
-                    .insert(id, top_level_comments.drain(..).collect());
-                self.metadata.insert(id, comments.drain(..).collect());
-                first = false;
-            }
-
-            scratch.insert(
-                source.into(),
-                TranslationEntry {
-                    // The three leading slots are positional metadata
-                    // (name/order/display name) and are only ever filled for the
-                    // first entry of a section, which consumes them above. For
-                    // every other entry they are empty placeholders that
-                    // `push_entries` would skip anyway, so dropping them here
-                    // avoids allocating a Vec of empty strings per entry.
-                    comments: replace(
-                        &mut comments,
-                        smallvec![String::new(); 3],
-                    )
-                    .into_iter()
-                    .filter(|comment| !comment.is_empty())
-                    .collect(),
-                    translation: translation.into(),
-                },
-            );
-        }
-
-        // Flush the last parsed section at EOF.
-        // Without this, the final `<!-- ID --><#>...` block is dropped if there
-        // is no following ID marker to trigger the regular section flush path.
-        if id != 0 {
-            let mut skip_entry = false;
-
-            if scratch.is_empty() {
-                let metadata_entry = self.metadata.entry(id).or_insert(
-                    replace(&mut comments, smallvec![String::new(); 3]),
-                );
-
-                let display_name = &metadata_entry[DISPLAY_NAME_POS];
-
-                if self.mode.is_write()
-                    && (display_name.is_empty()
-                        || display_name.ends_with(SEPARATOR))
-                {
-                    skip_entry = true;
-                } else {
-                    self.translation_maps
-                        .entry(id)
-                        .or_insert(TranslationMap::with_capacity(512));
-                }
-            }
-
-            if !skip_entry {
-                self.translation_maps
-                    .insert(id, scratch.drain(..).collect());
-            }
-        }
-
-        self.build_write_lookup();
-
-        Ok(())
-    }
-
-    /// Flattens `self.translation_maps` into `self.write_lookup`.
-    ///
-    /// Only relevant on write with [`DuplicateMode::Remove`], where [`Base::get_key`]
-    /// has to resolve a key against *every* parsed map. Doing that by scanning is
-    /// `O(maps)` per lookup, i.e. quadratic over the whole file set; flattening once
-    /// makes it a single hash lookup.
-    ///
-    /// Entries are moved rather than cloned, so this costs no extra memory: with
-    /// [`DuplicateMode::Remove`] the per-id maps are only consulted for presence on
-    /// write, never for content. `u16::MAX` is left alone - it holds the Termina item
-    /// category map, which is looked up by id directly.
-    fn build_write_lookup(&mut self) {
-        if !self.mode.is_write() || !self.duplicate_mode.is_remove() {
-            return;
-        }
-
-        let total: usize = self
-            .translation_maps
-            .iter()
-            .filter(|(id, _)| **id != u16::MAX)
-            .map(|(_, map)| map.len())
-            .sum();
-
-        self.write_lookup = TranslationMap::with_capacity(total);
-
-        for (id, map) in &mut self.translation_maps {
-            if *id == u16::MAX {
-                continue;
-            }
-
-            // `or_insert` keeps the first occurrence, matching the order the
-            // previous linear scan resolved duplicates in.
-            for (source, translation) in map.drain(..) {
-                self.write_lookup.entry(source).or_insert(translation);
-            }
-        }
-    }
-
-    /// The currently selected translation map.
-    ///
-    /// # Panics
-    ///
-    /// Panics if no map is selected, i.e. if called before [`Base::get_translation_map`].
-    fn translation_map(&self) -> &TranslationMap {
-        &self.translation_maps[self.translation_map_index]
-    }
-
-    /// Mutable counterpart of [`Base::translation_map`].
-    ///
-    /// # Panics
-    ///
-    /// Panics if no map is selected, i.e. if called before [`Base::get_translation_map`].
-    fn translation_map_mut(&mut self) -> &mut TranslationMap {
-        &mut self.translation_maps[self.translation_map_index]
-    }
-
-    /// Sets `self.translation_map` to the entry from `self.translation_maps`.
-    ///
-    /// If `self.mode` is [`Mode::Purge`], it will push entries from `self.translation_map` to `self.accumulated_translation` and break.
-    ///
-    /// If `self.flags` contains any of ignore flags, it will also set `self.ignore_entry`.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - ID of the entry to get.
-    ///
-    /// # Returns
-    ///
-    /// - [`ControlFlow::Break`]
-    ///     - If mode is [`Mode::Write`] and `id` is not in `self.translation_maps`.
-    ///     - If `id` is skipped.
-    ///     - If mode is [`Mode::Purge`].
-    /// - [`ControlFlow::Continue`] - In other situations.
-    ///
-    fn get_translation_map(&mut self, id: u16) -> ControlFlow<()> {
-        let entry = self.translation_maps.entry(id);
-        let index = entry.index();
-
-        // Select the map for `id`, creating it if we're not writing.
-        match entry {
-            Entry::Occupied(_) => {}
-            Entry::Vacant(entry) => {
-                if self.mode.is_write() {
-                    return ControlFlow::Break(());
-                }
-
-                entry.insert(TranslationMap::with_capacity(512));
-            }
-        }
-
-        self.translation_map_index = index;
-
-        if self
-            .skip_events
-            .get(&self.file_type)
-            .is_some_and(|x| x.contains(&id))
-            || (self.file_type.is_map() && self.skip_maps.contains(&id))
-        {
-            if self.mode.is_append() || self.mode.is_purge() {
-                let metadata = self.get_metadata(id);
-
-                let map = take(self.translation_map_mut());
-                self.accumulated_translation.push((
-                    id,
-                    metadata,
-                    FlushedLines::EMPTY,
-                    map,
-                ));
-            }
-
-            self.total_length = self.lines.len();
-            return ControlFlow::Break(());
-        }
-
-        self.get_ignore_entry(id);
-
-        if self.mode.is_purge() {
-            self.flush_translation(id);
-            return ControlFlow::Break(());
-        }
-
-        ControlFlow::Continue(())
-    }
-
-    /// Wraps string in a [`Value`].
-    ///
-    /// If `literal` argument is set, this wraps string in a [`Value`] of [`ValueType::String`] type.
-    /// Else, this wraps string in a [`Value`] of [`ValueType::Bytes`] type.
-    ///
-    /// # Parameters
-    ///
-    /// - `string` - String to wrap in a [`Value`].
-    /// - `literal` - Whether to wrap `string` as [`ValueType::String`] or as [`ValueType::Bytes`].
-    ///
-    /// # Returns
-    ///
-    /// - [`Value`] - wrapped string.
-    ///
-    fn make_string_value(string: &str, literal: bool) -> Value {
-        if literal {
-            Value::string(string)
-        } else {
-            Value::bytes(string.as_bytes())
-        }
-    }
-
-    fn string_is_only_symbols(string: &str) -> bool {
-        !string.chars().any(|c| !SYMBOLS.contains(&c))
-    }
-
-    // TODO(v15): Check when starts with if
-    //* This is breaking
-    fn ends_with_if_index(string: &str) -> Option<usize> {
-        if !string.ends_with(')') {
-            return None;
-        }
-
-        let mut stage: u8 = 0;
-        let char_indices = string.char_indices().rev().skip(1);
-
-        for (i, char) in char_indices {
-            match stage {
-                0 => {
-                    if char == '(' {
-                        stage = 1;
-                    }
-                }
-                1 => {
-                    if char == 'f' {
-                        stage = 2;
-                    } else {
-                        return None;
-                    }
-                }
-                2 => {
-                    if char == 'i' {
-                        return Some(i);
-                    }
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        None
-    }
-
-    fn find_lisa_prefix_index(string: &str) -> Option<usize> {
-        if string.starts_with(r"\et[") {
-            let mut index = r"\et[".len() + 1;
-
-            loop {
-                let char = string.as_bytes()[index];
-
-                if char == b']' {
-                    return Some(index + 1);
-                }
-
-                index += 1;
-
-                if index == 10 {
-                    return None;
-                }
-            }
-        } else if string.starts_with(r"\nbt") {
-            Some(r"\nbt".len())
-        } else if string.starts_with(r"\nblt") {
-            Some(r"\nblt".len())
-        } else {
-            None
-        }
-    }
-
-    /// Extracts string from [`Value`].
-    ///
-    /// Will always return [`None`] if [`Value`] is not of [`ValueType::String`] or [`ValueType::Bytes`].
-    ///
-    /// # Parameters
-    ///
-    /// - `value` - Value from which string will be extracted.
-    /// - `fail_if_empty` - Whether to return if extracted string happens to be empty.
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if [`Value`] is not of [`ValueType::String`] or [`ValueType::Bytes`], or `fail_if_empty` is set and `string` is empty.
-    /// - [`&str`] - Parsed string.
-    ///
-    /// The returned string borrows `value`, not `self` - keeping the lifetimes
-    /// separate lets the caller drop the `&self` borrow immediately.
-    fn extract_string<'v>(
-        &self,
-        value: &'v Value,
-        fail_if_empty: bool,
-    ) -> Option<&'v str> {
-        let string = value.as_str().or_else(|| {
-            std::str::from_utf8(value.as_byte_vec().unwrap_or_default()).ok()
-        })?;
-
-        let trimmed = string.trim();
-
-        if trimmed.is_empty() && fail_if_empty {
-            return None;
-        }
-
-        Some(if self.flags.contains(BaseFlags::Trim) {
-            trimmed
-        } else {
-            string
-        })
-    }
-
-    /// ONLY CALLED ON WRITE.
-    ///
-    /// Gets the [`TranslationEntry`] corresponding to the `key` from translation.
-    ///
-    /// This will return [`TranslationEntry`] corresponding to the `key` from `self.translation_map`, and also will seek it in maps `self.translation_maps` if `self.duplicate_mode` is [`DuplicateMode::Remove`].
-    ///
-    /// # Parameters
-    ///
-    /// - `key` - key to get.
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if key wasn't found in translation.
-    /// - [`&TranslationEntry`] - entry corresponding to the `key`.
-    ///
-    fn get_key(&self, key: &str) -> Option<&TranslationEntry> {
-        if self.duplicate_mode.is_allow() {
-            self.translation_map().get(key)
-        } else {
-            self.write_lookup.get(key)
-        }
-    }
-
-    /// Returns the RPG Maker data if `self.mode` is [`Mode::Write`], else returns translation data.
-    ///
-    /// # Parameters
-    ///
-    /// - `value` - [`Value`] to use on write.
-    ///
-    /// # Returns
-    ///
-    /// - [`ProcessedData::RPGMData`] if `self.mode` is [`Mode::Write`].
-    /// - [`ProcessedData::TranslationData`] otherwise.
-    ///
-    fn finish(&mut self, value: Value) -> ProcessedData {
-        if self.mode.is_write() {
-            ProcessedData::RPGMData(if self.file_type.is_plugins() {
-                let plugins_bytes = unsafe {
-                    to_vec(&SerdeValue::from(value)).unwrap_unchecked()
-                };
-
-                ["var $plugins =\n".as_bytes(), &plugins_bytes].concat()
-            } else if self.engine_type.is_new() {
-                unsafe { to_vec(&SerdeValue::from(value)).unwrap_unchecked() }
-            } else {
-                dump(
-                    value,
-                    if self.file_type.is_scripts() {
-                        None
-                    } else {
-                        INSTANCE_VAR_PREFIX
-                    },
-                )
-            })
-        } else {
-            self.finish_translation()
-        }
-    }
-
-    fn get_metadata(&mut self, id: u16) -> Comments {
-        let Some(mut comments) = self.metadata.remove(&id) else {
-            return SmallVec::default();
-        };
-
-        comments.iter_mut().enumerate().filter(|(_, x)| !x.is_empty()).for_each(|(i, x)| {
-            let pos = unsafe { transmute::<i8, CommentPos>(i as i8) };
-
-            *x = match pos {
-                CommentPos::Name => {
-                    format!("{NAME_COMMENT}{SEPARATOR}{x}")
-                }
-
-                CommentPos::Order => {
-                    format!("{MAP_ORDER_COMMENT}{SEPARATOR}{x}")
-                }
-
-                CommentPos::DisplayName => {
-                    let (source, translation) = x.split_once(SEPARATOR).unwrap();
-                    format!("{MAP_DISPLAY_NAME_COMMENT_PREFIX}{source}{COMMENT_SUFFIX}{SEPARATOR}{translation}")
-                }
-
-                CommentPos::None => unreachable!()
-            }
-        });
-
-        comments
-    }
-
-    fn finish_translation(&mut self) -> ProcessedData {
-        let allow_dup =
-            self.duplicate_mode.is_allow() || self.file_type.is_misc();
-        let skip_events_entry = self.skip_events.get(&self.file_type);
-        let ignore_index = self.ignore_entry_index;
-
-        let additional_data = self.get_additional_data();
-
-        // Allocate 4 MB. It makes no sense to circlejerk `accumulated_translation` to get the precise count, so we'll just take the biggest reasonable amount.
-        let output_size = 4096 * 1024;
-        let mut output = Vec::with_capacity(output_size);
-
-        for &data in additional_data {
-            output.extend_from_slice(data.as_bytes());
-            output.extend_from_slice(SEPARATOR.as_bytes());
-
-            if let Some(additional) = self.translation_maps.get(&u16::MAX) {
-                if let Some(translation) = additional.get(data) {
-                    output.extend_from_slice(translation.as_bytes());
-                }
-            }
-
-            output.push(b'\n');
-        }
-
-        let mut accumulated_map: indexmap::IndexMap<
-            String,
-            (u16, TranslationEntry),
-            gxhash::GxBuildHasher,
-        > = if allow_dup {
-            indexmap::IndexMap::default()
-        } else {
-            let len = self.translation_maps.values().fold(0, |mut acc, map| {
-                acc += map.len();
-                acc
-            });
-
-            self.translation_maps.drain(..).fold(
-                indexmap::IndexMap::with_capacity_and_hasher(
-                    len,
-                    GxBuildHasher::default(),
-                ),
-                |mut acc, (k, v)| {
-                    for (key, value) in v {
-                        acc.insert(key, (k, value));
-                    }
-                    acc
-                },
-            )
-        };
-
-        // Moved out so the loop can read the rest of `self` freely. `lines_store`
-        // backs every `FlushedLines::Range`.
-        let mut accumulated = take(&mut self.accumulated_translation);
-        let lines_store = take(&mut self.lines);
-        let mut prev_id = u16::MAX;
-
-        for i in 0..accumulated.len() {
-            // Splitting gives the current entry mutably and the lookahead
-            // immutably at the same time.
-            let (current, rest) = accumulated[i..].split_first_mut().unwrap();
-            let (id, meta, lines, map) = current;
-
-            let skip = skip_events_entry.is_some_and(|e| e.contains(id))
-                || (self.file_type.is_map() && self.skip_maps.contains(id))
-                || (self.mode.is_purge()
-                    && self.file_type.is_system()
-                    && *id == 8);
-
-            if skip {
-                if self.mode.is_append() || self.mode.is_purge() {
-                    push_metadata(&mut output, *id, meta);
-
-                    for (source, translation) in map {
-                        push_entries(&mut output, source, translation);
-                    }
-                }
-
-                continue;
-            }
-
-            if let Some(comments) = self.top_level_comments.get(id) {
-                for comment in comments {
-                    output.extend_from_slice(comment.as_bytes());
-                    output.push(b'\n');
-                }
-            }
-
-            if self.mode.is_purge() {
-                push_metadata(&mut output, *id, meta);
-
-                for (mut source, translation) in take(map) {
-                    if translation.is_empty() {
-                        let moved = take(&mut source);
-
-                        if self.flags.contains(BaseFlags::CreateIgnore)
-                            && !moved.is_empty()
-                        {
-                            // Field access rather than `ignore_entry_mut`, so
-                            // that `skip_events_entry`'s borrow of the disjoint
-                            // `skip_events` field stays live.
-                            if let Some((_, entry)) =
-                                self.ignore_map.get_index_mut(ignore_index)
-                            {
-                                entry.insert(moved);
-                            }
-                        }
-                    }
-
-                    push_entries(&mut output, &source, &translation);
-                }
-
-                continue;
-            }
-
-            if *id != prev_id {
-                let has_display_name =
-                    meta.get(DISPLAY_NAME_POS).is_some_and(|c| !c.is_empty());
-
-                let same_id_has_lines = !lines.is_empty()
-                    || rest.iter().any(|(next_id, _, next_lines, _)| {
-                        *next_id == *id && !next_lines.is_empty()
-                    });
-
-                let should_push_map = self.file_type.is_map()
-                    && self.map_events
-                    && (same_id_has_lines || has_display_name);
-
-                let should_push_other = !lines.is_empty() || has_display_name;
-
-                if should_push_map || should_push_other {
-                    push_metadata(&mut output, *id, meta);
-                }
-
-                prev_id = *id;
-            }
-
-            let next_lines_empty = rest
-                .first()
-                .is_some_and(|(_, _, next_lines, _)| next_lines.is_empty());
-
-            if !next_lines_empty {
-                if let Some((_, entry)) = map.first() {
-                    push_entries(&mut output, "", entry);
-                }
-            }
-
-            for line_index in 0..lines.len() {
-                let source = lines.get(&lines_store, line_index);
-
-                let translation = match (allow_dup, self.mode.is_append()) {
-                    (true, true) => {
-                        map.swap_remove(source).unwrap_or_default()
-                    }
-                    (false, true) => accumulated_map
-                        .swap_remove(source)
-                        .unzip()
-                        .1
-                        .unwrap_or_default(),
-                    (_, false) => TranslationEntry::default(),
-                };
-
-                push_entries(&mut output, source, &translation);
-            }
-
-            if self.flags.contains(BaseFlags::SkipObsolete) {
-                continue;
-            }
-
-            match (allow_dup, self.mode.is_append()) {
-                (true, true) => {
-                    for (source, translation) in map {
-                        push_entries(&mut output, source, translation);
-                    }
-                }
-                (false, true) => {
-                    for (source, (i, translation)) in &accumulated_map {
-                        if *id == *i {
-                            push_entries(&mut output, source, translation);
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        output.pop();
-        ProcessedData::TranslationData(output)
-    }
-
-    /// Flushes current `self.translation_map` and `self.lines` to `self.accumulated_translation` along with metadata and id.
-    ///
-    /// It's necessary to call [`Base::finish_translation`] once we've finished flushing entries.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - ID of the entry to flush.
-    ///
-    fn flush_translation(&mut self, id: u16) {
-        let metadata = self.get_metadata(id);
-
-        if self.mode.is_purge() {
-            if !self.translation_map().is_empty()
-                || metadata
-                    .get(DISPLAY_NAME_POS)
-                    .is_some_and(|x| !x.is_empty())
-            {
-                let map = take(self.translation_map_mut());
-                self.accumulated_translation.push((
-                    id,
-                    metadata,
-                    FlushedLines::EMPTY,
-                    map,
-                ));
-            }
-        } else if self.duplicate_mode.is_allow() || self.file_type.is_misc() {
-            if self
-                .skip_events
-                .get(&self.file_type)
-                .is_some_and(|e| e.contains(&id))
-                || (self.file_type.is_map() && self.skip_maps.contains(&id))
-            {
-                self.lines.clear();
-                self.translation_map_mut().clear();
-            } else {
-                let lines =
-                    FlushedLines::Owned(self.lines.drain(..).collect());
-
-                let map = take(self.translation_map_mut());
-                self.accumulated_translation.push((
-                    id,
-                    metadata,
-                    lines,
-                    map,
-                ));
-            }
-        } else {
-            let total_length = self.total_length;
-            let current_length = self.lines.len() - total_length;
-
-            // Left in `self.lines`; only the range is recorded, so nothing is copied.
-            let lines =
-                FlushedLines::Range(total_length..total_length + current_length);
-
-            self.accumulated_translation.push((
-                id,
-                metadata,
-                lines,
-                TranslationMap::default(),
-            ));
-
-            self.total_length += current_length;
-        }
-    }
-
-    fn update_metadata(
-        &mut self,
-        id: u16,
-        metadata_vec: Vec<(CommentPos, &str)>,
-    ) {
-        let metadata = self
-            .metadata
-            .entry(id)
-            .or_insert(smallvec![String::new(); 3]);
-
-        if metadata.len() < 3 {
-            metadata.resize(3, String::new());
-        }
-
-        for (entry_id, entry) in
-            metadata_vec.into_iter().filter(|(_, x)| !x.is_empty())
-        {
-            if entry_id == CommentPos::DisplayName {
-                if self.mode.is_append() {
-                    let Some((source, mut translation)) =
-                        metadata[entry_id as usize].split_once(SEPARATOR)
-                    else {
-                        metadata[entry_id as usize] =
-                            format!("{entry}{SEPARATOR}");
-                        continue;
-                    };
-
-                    if source != entry {
-                        translation = "";
-                    }
-
-                    metadata[entry_id as usize] =
-                        format!("{entry}{SEPARATOR}{translation}");
-                } else {
-                    metadata[entry_id as usize] = format!("{entry}{SEPARATOR}");
-                }
-
-                continue;
-            }
-
-            metadata[entry_id as usize] = entry.to_string();
-        }
-    }
-
-    /// Returns some additional data, that needs to be inserted at the start of the output translation data.
-    ///
-    /// Right now, this function returns the slice of source entries that need to be inserted on read. This may change at any moment.
-    ///
-    /// # Returns
-    ///
-    /// - [`&[&str]`] - slice of source entries.
-    ///
-    #[must_use]
-    fn get_additional_data(&self) -> &[&str] {
-        if !self.mode.is_write()
-            && self.game_type.is_termina()
-            && self.file_type.is_items()
-        {
-            return &[
-                "<Menu Category: Items>",
-                "<Menu Category: Food>",
-                "<Menu Category: Healing>",
-                "<Menu Category: Body bag>",
-            ];
-        }
-
-        &[]
-    }
-}
-
-/// Base for processing `Map` files.
-pub struct MapBase<'a> {
-    pub base: &'a mut Base,
-    mapinfos: Value,
-}
-
-impl<'a> MapBase<'a> {
-    /// Initializes system base using [`Base`].
-    /// Before calling this, you should create a base and pass it here.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use rvpacker_txt_rs_lib::{core::{Base, MapBase}, Mode, ReadMode, EngineType};
-    ///
-    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    /// let mut map_base = MapBase::new(&mut base);
-    /// ```
-    pub fn new(base: &'a mut Base) -> Self {
-        base.reset();
-        base.file_type = RPGMFileType::Map;
-
-        Self {
-            base,
-            mapinfos: Value::default(),
-        }
-    }
-
-    /// Returns the translation data, accumulated after processing multiple maps.
-    ///
-    /// Returns the actual data only with [`Mode::Read`] or [`Mode::Purge`].
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, MapBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    ///     let mut map_base = MapBase::new(&mut base);
-    ///
-    ///     let mapinfos = read("C:/Game/Data/MapInfos.rvdata2")?;
-    ///
-    ///     let map_file_content = read("C:/Game/Data/Map001.rvdata2")?;
-    ///     let data = map_base.process("Map001.rvdata2", &map_file_content, &mapinfos, None)?;
-    ///
-    ///     let map_file_content = read("C:/Game/Data/Map002.rvdata2")?;
-    ///     let data = map_base.process("Map002.rvdata2", &map_file_content, &mapinfos, None)?;
-    ///
-    ///     let translation_data = map_base.translation();
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn translation(&mut self) -> ProcessedData {
-        self.base.finish(Value::default())
-    }
-
-    /// Processes the RPG Maker map file content.
-    ///
-    /// To get the translation data, you need to call [`MapBase::translation`] after processing required maps.
-    ///
-    /// # Parameters
-    ///
-    /// - `filename` - Filename of the file that's being processed.
-    /// - `content` - Content of the file that's being processed.
-    /// - `mapinfos` - `MapInfos` file content that corresponds to the file being parsed.
-    /// - `translation` - Contents of the translation file corresponding to maps. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if map is unused (not included in Mapinfos), or mode is [`Mode::Write`] and no translation exists for the map.
-    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
-    /// - [`Error`], if unable to parse the content.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::MarshalLoad`] - if unable to load the Marshal data.
-    /// - [`Error::JsonParse`] - if unable to parse the JSON data.
-    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
-    ///
-    /// # Panics
-    ///
-    /// May panic if passed content is not from `Map` file.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, MapBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    ///     let mut map_base = MapBase::new(&mut base);
-    ///
-    ///     let map_file_content = read("C:/Game/Data/Map001.rvdata2")?;
-    ///     let mapinfos = read("C:/Game/Data/MapInfos.rvdata2")?;
-    ///     let data = map_base.process("Map001.rvdata2", &map_file_content, &mapinfos, None)?;
-    ///
-    ///     // Required only when reading.
-    ///     let translation_data = map_base.translation();
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn process(
-        &mut self,
-        filename: &str,
-        content: &[u8],
-        mapinfos: &[u8],
-        translation: Option<&str>,
-    ) -> Result<Option<ProcessedData>, Error> {
-        if self.mapinfos.is_null() {
-            self.mapinfos = parse_rpgm_file(
-                mapinfos,
-                self.base.engine_type,
-                self.base.file_type,
-            )?;
-        }
-
-        self.base.initialize_translation(translation)?;
-
-        let id = Self::parse_map_id(filename);
-        if self.is_map_unused(id) {
-            return Ok(None);
-        }
-
-        if self.base.get_translation_map(id).is_break() {
-            return Ok(None);
-        }
-
-        let mut map_object = parse_rpgm_file(
-            content,
-            self.base.engine_type,
-            self.base.file_type,
-        )?;
-        let display_name = self.get_display_name(&map_object);
-
-        if self.base.mode.is_read() {
-            let map_order = self.get_map_order(id).to_string();
-            let engine_is_new = self.base.engine_type.is_new();
-            let map_name =
-                Self::get_map_name(&self.mapinfos, engine_is_new, id);
-            let replaced_map_name = map_name.normalize();
-
-            self.base.update_metadata(
-                id,
-                Vec::from([
-                    (CommentPos::Name, replaced_map_name.as_ref()),
-                    (CommentPos::Order, &map_order),
-                    (CommentPos::DisplayName, &display_name),
-                ]),
-            );
-        } else if !display_name.is_empty() {
-            let display_name_comment_line = &self.base.metadata[&id][2];
-
-            let split: Vec<&str> =
-                display_name_comment_line.split(SEPARATOR).collect();
-
-            if split.len() >= 2 {
-                let mut translation = split
-                    .into_iter()
-                    .skip(1)
-                    .rfind(|x| !x.is_empty())
-                    .unwrap_or_default();
-
-                let translation_replaced = translation.denormalize();
-                translation = &translation_replaced;
-
-                map_object[self.base.labels.display_name] =
-                    Value::string(translation);
-            } else {
-                log::warn!(
-                    "{COULD_NOT_SPLIT_LINE_MSG} \
-                     {display_name_comment_line}\n{IN_FILE_MSG}: {file}.txt",
-                    file = self.base.file_type.to_string().to_lowercase()
-                );
-            }
-        }
-
-        let events = if self.base.engine_type.is_new() {
-            // Previously, this line was using `unwrap_unchecked`, because it assumed, that events are always an array in MV/MZ.
-            // This is not the case. This array can also contain just `bool`. Now, it returns, if encounters something else than an array.
-            let Some(array) =
-                map_object[self.base.labels.events].as_array_mut()
-            else {
-                return Ok(None);
-            };
-
-            EventIterator::New(array.iter_mut().skip(1))
-        } else {
-            // SAFETY: Always a hashmap in old maps.
-            EventIterator::Old(unsafe {
-                map_object[self.base.labels.events]
-                    .as_hashmap_mut()
-                    .unwrap_unchecked()
-                    .values_mut()
-            })
-        };
-
-        for event in events {
-            if event.is_null() {
-                continue;
-            }
-
-            // Read before borrowing `pages` out of the same event.
-            let event_metadata = if self.base.map_events {
-                Some((
-                    event["id"].as_int().unwrap(),
-                    event["name"].as_str().unwrap().to_owned(),
-                    event["x"].as_int().unwrap(),
-                    event["y"].as_int().unwrap(),
-                ))
-            } else {
-                None
-            };
-
-            let Some(pages) =
-                event[self.base.labels.pages].as_array_mut()
-            else {
-                continue;
-            };
-
-            if let Some((event_id, event_name, event_x, event_y)) =
-                event_metadata
-            {
-                self.base.flush_translation(id);
-
-                self.base.accumulated_translation.push((
-                        id,
-                        SmallVec::default(),
-                        FlushedLines::EMPTY,
-                        TranslationMap::from_iter([(String::new(), TranslationEntry {
-                            comments: vec![format!(
-                                "{EVENT_ID_COMMENT}{SEPARATOR}{event_id}"
-                            ),
-                            format!("{EVENT_NAME_COMMENT}{SEPARATOR}{event_name}"),
-                            format!("{EVENT_POS_COMMENT}{SEPARATOR}{event_x},{event_y}")],
-                            translation: String::new(),
-                        })])
-                    ));
-            }
-
-            for page in pages {
-                // SAFETY: List is always in map files.
-                let list = unsafe {
-                    page[self.base.labels.list]
-                        .as_array_mut()
-                        .unwrap_unchecked()
-                };
-
-                self.base.process_list(list);
-            }
-        }
-
-        if self.base.mode.is_write() {
-            Ok(Some(self.base.finish(map_object)))
-        } else {
-            self.base.flush_translation(id);
-            Ok(None)
-        }
-    }
-
-    /// Parses a map ID from a filename by extracting digits starting from position 3 and parsing them to [`u16`].
-    ///
-    /// # Parameters
-    ///
-    /// - `filename` - Filename of the map.
-    ///
-    /// # Returns
-    ///
-    /// - [`u16`] - The parsed map ID.
-    ///
-    pub fn parse_map_id(filename: &str) -> u16 {
-        let mut id: u16 = 0;
-
-        // Accumulated directly rather than copied into a buffer and re-parsed.
-        // Not capped at three digits, because a game may have more than 999 maps.
-        for &byte in &filename.as_bytes()[3..] {
-            if !byte.is_ascii_digit() {
-                break;
-            }
-
-            id = id * 10 + u16::from(byte - b'0');
-        }
-
-        id
-    }
-
-    /// Determines whether a map is unused based on its existence in `self.mapinfos`.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - The ID of the map to check.
-    ///
-    /// # Returns
-    ///
-    /// - [`bool`] - Whether map is unused.
-    ///
-    fn is_map_unused(&self, id: u16) -> bool {
-        // If map ID can't be found in mapinfos, then it is unused in game.
-        if self.base.engine_type.is_new() {
-            self.mapinfos.get_index(id as usize).is_none()
-        } else {
-            self.mapinfos.get(&Value::int(i32::from(id))).is_none()
-        }
-    }
-
-    /// Retrieves the chronological map order from `self.mapinfos`.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - The ID of the map whose order should be retrieved.
-    ///
-    /// # Returns
-    ///
-    /// - [`u16`] - The map's order.
-    ///
-    fn get_map_order(&self, id: u16) -> i32 {
-        // SAFETY: "order" always exists in mapinfos and is always an integer.
-        unsafe {
-            if self.base.engine_type.is_new() {
-                &self.mapinfos[id as usize]["order"]
-            } else {
-                &self.mapinfos[Value::int(i32::from(id))]["order"]
-            }
-            .as_int()
-            .unwrap_unchecked()
-        }
-    }
-
-    /// Retrieves the name of the map as a string slice, based on the provided map ID.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` - The ID of the map whose name should be retrieved.
-    ///
-    /// # Returns
-    ///
-    /// - [`&str`] - The name of the map.
-    ///
-    /// Takes `mapinfos` explicitly rather than reading it through `&self`, so the
-    /// returned string borrows only that field and leaves `self.base` free.
-    fn get_map_name(mapinfos: &Value, engine_is_new: bool, id: u16) -> &str {
-        // SAFETY: "name" always exists in mapinfos and is always a string.
-        unsafe {
-            if engine_is_new {
-                &mapinfos[id as usize]["name"]
-            } else {
-                &mapinfos[Value::int(i32::from(id))]["name"]
-            }
-            .as_str()
-            .unwrap_unchecked()
-        }
-    }
-
-    /// Retrieves a display name for a map object.
-    ///
-    /// # Parameters
-    ///
-    /// - `map_object` - A reference to a [`Value`] representing the map object.
-    ///
-    /// # Returns
-    ///
-    /// - [`String`] - The processed display name, or an empty string if not found.
-    ///
-    fn get_display_name(&self, map_object: &Value) -> String {
-        map_object
-            .get(self.base.labels.display_name)
-            .map(|display_name| {
-                display_name
-                    .as_str()
-                    .map(|name| name.normalize().into_owned())
-                    .unwrap_or_default()
-            })
-            .unwrap_or_default()
-    }
-}
-
-/// Base for processing other files (`Actors`, `Armors`, `Classes`, `Enemies`, `CommonEvents`, `Troops`, `Items`, `Skills`, `States`, `Weapons`).
-pub struct OtherBase<'a> {
-    pub base: &'a mut Base,
-}
-
-impl<'a> OtherBase<'a> {
-    /// Initializes system base using [`Base`].
-    /// Before calling this, you should create a base and pass it here.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use rvpacker_txt_rs_lib::{core::{Base, OtherBase}, Mode, ReadMode, EngineType};
-    ///
-    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    /// let mut other_base = OtherBase::new(&mut base);
-    /// ```
-    pub fn new(base: &'a mut Base) -> Self {
-        base.reset();
-        base.file_type = RPGMFileType::Invalid;
-
-        Self { base }
-    }
-
-    /// Processes the RPG Maker other file content.
-    ///
-    /// # Parameters
-    ///
-    /// - `filename` - Filename of the file that's being processed.
-    /// - `content` - Content of the file that's being processed.
-    /// - `translation` - Contents of the translation file corresponding to the file. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if `mode` is [`Mode::Write`] and no translation exists.
-    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
-    /// - [`Error`], if unable to parse the content.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::MarshalLoad`] - if unable to load the Marshal data.
-    /// - [`Error::JsonParse`] - if unable to parse the JSON data.
-    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
-    ///
-    /// # Panics
-    ///
-    /// May panic if passed content is not `Actors`, `Armors`, `Classes`, `Enemies`, `CommonEvents`, `Troops`, `Items`, `Skills`, `States`, `Weapons`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, OtherBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    ///     let mut other_base = OtherBase::new(&mut base);
-    ///
-    ///     let other_file_content = read("C:/Game/Data/Actors.rvdata2")?;
-    ///     other_base.process("Actors.rvdata2", &other_file_content, None)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn process(
-        &mut self,
-        filename: &str,
-        content: &[u8],
-        translation: Option<&str>,
-    ) -> Result<Option<ProcessedData>, Error> {
-        self.base.file_type = RPGMFileType::from_filename(filename);
-
-        self.base.reset();
-        self.base.initialize_translation(translation)?;
-
-        let mut entry_value = parse_rpgm_file(
-            content,
-            self.base.engine_type,
-            self.base.file_type,
-        )?;
-
-        // SAFETY: All "other" entries are always arrays.
-        let object_array =
-            unsafe { entry_value.as_array_mut().unwrap_unchecked() };
-
-        let mut processed = false;
-
-        // Skipping one, because the first entry is always null.
-        for object in object_array.iter_mut().skip(1) {
-            // SAFETY: Name and ID exists on every object.
-            let id = unsafe { object["id"].as_int().unwrap_unchecked() } as u16;
-
-            if self.base.get_translation_map(id).is_break() {
-                if self.base.mode.is_purge() {
-                    processed = true;
-                }
-
-                continue;
-            }
-
-            processed = true;
-
-            let event_name = unsafe {
-                object[self.base.labels.name].as_str().unwrap_unchecked()
-            };
-
-            self.base.update_metadata(
-                id,
-                Vec::from([(CommentPos::Name, event_name)]),
-            );
-
-            if self.base.file_type.is_events()
-                || self.base.file_type.is_troops()
-            {
-                self.process_object(object);
-            } else {
-                self.process_array(object);
-            }
-
-            self.base.flush_translation(id);
-        }
-
-        if !processed {
-            return Ok(None);
-        }
-
-        Ok(Some(self.base.finish(entry_value)))
-    }
-
-    fn process_variable_termina(
-        &self,
-        mut variable_text: Cow<'_, str>,
-        variable_type: Variable,
-        note_text: Option<&str>,
-    ) -> Option<String> {
-        if variable_text.starts_with("///") || variable_text.contains("---") {
-            return None;
-        }
-
-        match variable_type {
-            Variable::Description => {
-                if let Some(note) = note_text {
-                    let mut note_is_continuation = false;
-
-                    if !note.starts_with("flesh puppetry") {
-                        let mut note_chars = note.chars();
-
-                        if let Some((note_first_char, note_second_char)) =
-                            note_chars.next().zip(note_chars.next())
-                        {
-                            let is_continuation = note_first_char == '\n'
-                                && note_second_char != '\n';
-
-                            let first_char_is_valid = note_first_char
-                                .is_ascii_alphabetic()
-                                || note_first_char == '"'
-                                || note.starts_with("4 sticks");
-
-                            let first_char_is_punctuation = matches!(
-                                note_first_char,
-                                '.' | '!' | '/' | '?'
-                            );
-
-                            if (is_continuation || first_char_is_valid)
-                                && !first_char_is_punctuation
-                            {
-                                note_is_continuation = true;
-                            }
-                        }
-                    }
-
-                    if note_is_continuation {
-                        let mut note_string = String::from(note);
-
-                        if let Some((mut left, _)) =
-                            note.trim_start().split_once('\n')
-                        {
-                            left = left.trim();
-
-                            if left.ends_with(['.', '%', '!', '"']) {
-                                note_string = String::from(
-                                    if self.base.mode.is_write() {
-                                        "\n"
-                                    } else {
-                                        NEW_LINE
-                                    },
-                                ) + left;
-                            } else if self.base.mode.is_read() {
-                                return None;
-                            }
-                        } else if note.ends_with(['.', '%', '!', '"'])
-                            || note.ends_with("takes place?")
-                        {
-                            note_string = note.into();
-                        } else if self.base.mode.is_read() {
-                            return None;
-                        }
-
-                        if note_string.is_empty() {
-                            if self.base.mode.is_read() {
-                                return None;
-                            }
-                        } else {
-                            variable_text = Cow::Owned(format!(
-                                "{variable_text}{note_string}"
-                            ));
-                        }
-                    }
-                }
-            }
-            Variable::Message1
-            | Variable::Message2
-            | Variable::Message3
-            | Variable::Message4 => {
-                return None;
-            }
-            Variable::Note => {
-                if self.base.mode.is_write() && self.base.file_type.is_items() {
-                    for string in [
-                        "<Menu Category: Items>",
-                        "<Menu Category: Food>",
-                        "<Menu Category: Healing>",
-                        "<Menu Category: Body bag>",
-                    ] {
-                        if variable_text.rfind(string).is_some() {
-                            return Some(variable_text.replace(
-                                string,
-                                &self.base.translation_maps[&u16::MAX][string],
-                            ));
-                        }
-                    }
-                }
-
-                if !self.base.file_type.is_classes() {
-                    return None;
-                }
-            }
-            Variable::Name | Variable::Nickname => match self.base.file_type {
-                RPGMFileType::Actors => {
-                    if ![
-                        "Levi",
-                        "Marina",
-                        "Daan",
-                        "Abella",
-                        "O'saa",
-                        "Blood golem",
-                        "Black Kalev",
-                        "Marcoh",
-                        "Karin",
-                        "Olivia",
-                        "Ghoul",
-                        "Villager",
-                        "August",
-                        "Caligura",
-                        "Henryk",
-                        "Pav",
-                        "Tanaka",
-                        "Samarie",
-                    ]
-                    .contains(&variable_text.as_ref())
-                    {
-                        return None;
-                    }
-                }
-                RPGMFileType::Armors => {
-                    if variable_text.starts_with("test_armor") {
-                        return None;
-                    }
-                }
-                RPGMFileType::Classes => {
-                    if [
-                        "Girl",
-                        "Kid demon",
-                        "Captain",
-                        "Marriage",
-                        "Marriage2",
-                        "Baby demon",
-                        "Buckman",
-                        "Nas'hrah",
-                        "Skeleton",
-                    ]
-                    .contains(&variable_text.as_ref())
-                    {
-                        return None;
-                    }
-                }
-                RPGMFileType::Enemies => {
-                    if ["Spank Tank", "giant", "test"]
-                        .contains(&variable_text.as_ref())
-                    {
-                        return None;
-                    }
-                }
-                RPGMFileType::Items => {
-                    if [
-                        "Torch",
-                        "Flashlight",
-                        "Stick",
-                        "Quill",
-                        "Empty scroll",
-                        "Soul stone_NOT_USE",
-                        "Cube of depths",
-                        "Worm juice",
-                        "Silver shilling",
-                        "Coded letter #1 - UNUSED",
-                        "Black vial",
-                        "Torturer's notes 1",
-                        "Purple vial",
-                        "Orange vial",
-                        "Red vial",
-                        "Green vial",
-                        "Pinecone pig instructions",
-                        "Grilled salmonsnake meat",
-                        "Empty scroll",
-                        "Water vial",
-                        "Blood vial",
-                        "Devil's Grass",
-                        "Stone",
-                        "Codex #1",
-                        "The Tale of the Pocketcat I",
-                        "The Tale of the Pocketcat II",
-                    ]
-                    .contains(&variable_text.as_ref())
-                        || variable_text.starts_with("The Fellowship")
-                        || variable_text.starts_with("Studies of")
-                        || variable_text.starts_with("Blueish")
-                        || variable_text.starts_with("Skeletal")
-                        || variable_text.ends_with("soul")
-                        || variable_text.ends_with("schematics")
-                    {
-                        return None;
-                    }
-                }
-                RPGMFileType::Weapons => {
-                    if variable_text == "makeshift2" {
-                        return None;
-                    }
-                }
-                _ => {}
-            },
-        }
-
-        Some(variable_text.into_owned())
-    }
-
-    #[allow(clippy::collapsible_match, clippy::single_match)]
-    fn process_variable(
-        &self,
-        variable_text: &str,
-        note_text: Option<&str>,
-        variable_type: Variable,
-    ) -> Option<String> {
-        if Base::string_is_only_symbols(variable_text) {
-            return None;
-        }
-
-        let mut variable_text = Cow::Borrowed(variable_text);
-
-        if !self.base.engine_type.is_new() {
-            if variable_text.lines().all(|line| {
-                line.is_empty()
-                    || IS_INVALID_MULTILINE_VARIABLE_RE
-                        .with(|r| r.is_match(line))
-            }) || IS_INVALID_VARIABLE_RE.with(|r| r.is_match(&variable_text))
-            {
-                return None;
-            }
-
-            variable_text = Cow::Owned(variable_text.replace("\r\n", "\n"));
-        }
-
-        match self.base.game_type {
-            GameType::Termina => {
-                if let Some(text) = self.process_variable_termina(
-                    variable_text,
-                    variable_type,
-                    note_text,
-                ) {
-                    if self.base.mode.is_write()
-                        && self.base.file_type.is_items()
-                        && variable_type.is_note()
-                    {
-                        return Some(text);
-                    }
-
-                    variable_text = Cow::Owned(text);
-                } else {
-                    return None;
-                }
-            }
-            // custom processing for other games
-            _ => {}
-        }
-
-        if self.base.mode.is_read() {
-            return Some(variable_text.into_owned());
-        }
-
-        let translated = self.base.get_key(&variable_text).map(|translated| {
-            let mut result = translated.to_string();
-
-            if variable_type.is_any_message()
-                && !(variable_type.is_message_2()
-                    && self.base.file_type.is_skills())
-            {
-                result = String::from(' ') + &result;
-            }
-
-            match self.base.game_type {
-                GameType::Termina => match variable_type {
-                    Variable::Note => {
-                        if let Some(first_char) = result.chars().next() {
-                            if first_char != '\n' {
-                                result = String::from('\n') + &result;
-                            }
-                        }
-                    }
-                    _ => {}
-                },
-                _ => {}
-            }
-
-            if self.base.game_type.is_termina()
-                && variable_type.is_description()
-            {
-                result += "\n\n\n\n";
-            }
-
-            result
-        });
-
-        translated
-    }
-
-    /// Processes an object from `CommonEvents` or `Troops` file.
-    fn process_object(&mut self, object: &mut Value) {
-        if self.base.file_type.is_troops() {
-            // SAFETY: Troops always include pages.
-            let pages = unsafe {
-                object[self.base.labels.pages]
-                    .as_array_mut()
-                    .unwrap_unchecked()
-            };
-
-            for page in pages {
-                if let Some(list_array) =
-                    page[self.base.labels.list].as_array_mut()
-                {
-                    self.base.process_list(list_array);
-                }
-            }
-        } else {
-            // SAFETY: CommonEvents always include list.
-            let list = unsafe {
-                object[self.base.labels.list]
-                    .as_array_mut()
-                    .unwrap_unchecked()
-            };
-
-            self.base.process_list(list);
-        }
-    }
-
-    /// Processes an object array from `Actors`, `Armors`, `Classes`, `Enemies`, `Items`, `States`, `Weapons` files.
-    fn process_array(&mut self, array: &mut Value) {
-        let variable_pairs = [
-            (self.base.labels.name, Variable::Name),
-            (self.base.labels.nickname, Variable::Nickname),
-            (self.base.labels.description, Variable::Description),
-            (self.base.labels.message1, Variable::Message1),
-            (self.base.labels.message2, Variable::Message2),
-            (self.base.labels.message3, Variable::Message3),
-            (self.base.labels.message4, Variable::Message4),
-            (self.base.labels.note, Variable::Note),
-        ];
-
-        for (variable_label, variable_type) in variable_pairs {
-            let Some(object) = array.get(variable_label) else {
-                continue;
-            };
-
-            let Some(string) = self.base.extract_string(object, true) else {
-                continue;
-            };
-
-            let mut string = Cow::Borrowed(string);
-
-            if self.base.mode.is_write() {
-                string = Cow::Owned(
-                    string
-                        .lines()
-                        .map(str::trim)
-                        .collect::<Vec<_>>()
-                        .join("\n"),
-                );
-            }
-
-            let note_text = if self.base.game_type.is_termina()
-                && variable_type.is_description()
-            {
-                array[self.base.labels.note].as_str()
-            } else {
-                None
-            };
-
-            let Some(parsed) =
-                self.process_variable(&string, note_text, variable_type)
-            else {
-                continue;
-            };
-
-            if self.base.mode.is_write() {
-                array[variable_label] = Value::string(parsed);
-            } else {
-                let folded = parsed.lines().fold(
-                    String::with_capacity(parsed.len() * 2),
-                    |mut output, line| {
-                        let trimmed = if variable_type.is_any_message()
-                            || self.base.flags.contains(BaseFlags::Trim)
-                        {
-                            line.trim()
-                        } else {
-                            line
-                        };
-
-                        let _ = write!(output, "{trimmed}{NEW_LINE}");
-
-                        output
-                    },
-                );
-
-                let replaced =
-                    unsafe { folded.strip_suffix(NEW_LINE).unwrap_unchecked() };
-
-                self.base.insert_string(Cow::Borrowed(replaced));
-            }
-        }
-    }
-}
-
-pub struct SystemBase<'a> {
-    pub base: &'a mut Base,
-    game_title: String,
-    system_value: Value,
-}
-
-impl<'a> SystemBase<'a> {
-    /// Initializes system base using [`Base`].
-    /// Before calling this, you should create a base and pass it here.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use rvpacker_txt_rs_lib::{core::{Base, SystemBase}, Mode, ReadMode, EngineType};
-    ///
-    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    /// let mut system_base = SystemBase::new(&mut base);
-    /// ```
-    pub fn new(base: &'a mut Base) -> Self {
-        base.reset();
-        base.file_type = RPGMFileType::System;
-
-        Self {
-            base,
-            game_title: String::new(),
-            system_value: Value::default(),
-        }
-    }
-
-    /// This function exists for compatibility with RPG Maker XP, VX and VX Ace. It should be called only when reading.
-    ///
-    /// RPG Maker XP/VX/VXA games may not contain game title in their respective system file. Instead, they may only contain the title in `Game.ini` file. This file is not necessarily UTF-8 encoded.
-    ///
-    /// Since there's no way to tell the encoding, it's user responsibility to call [`get_ini_title`], find title's encoding through trial-and-error, and pass it here.
-    ///
-    /// Passed title overrides automatic extraction; that means that passed title will be preferred over the title from the system file, if title even exists there.
-    ///
-    /// # Parameters
-    ///
-    /// `title` - UTF-8 encoded [`&str`] title.
-    ///
-    /// # Note
-    ///
-    /// This function is no-op if mode is not [`Mode::Read`].
-    ///
-    pub fn set_game_title(&mut self, title: &str) {
-        if self.base.mode.is_read() {
-            self.game_title = title.to_string();
-        }
-    }
-
-    /// Processes the RPG Maker system file content.
-    ///
-    /// # Parameters
-    ///
-    /// - `content` - Content of the file that's being processed.
-    /// - `translation` - Contents of the translation file corresponding to the file. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if `mode` is [`Mode::Write`] and no translation exists.
-    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
-    /// - [`Error`], if unable to parse the content.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::MarshalLoad`] - if unable to load the Marshal data.
-    /// - [`Error::JsonParse`] - if unable to parse the JSON data.
-    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
-    ///
-    /// # Panics
-    ///
-    /// May panic if passed content is not `System`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, SystemBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    ///     let mut system_base = SystemBase::new(&mut base);
-    ///
-    ///     let system_file_content = read("C:/Game/Data/System.rvdata2")?;
-    ///     system_base.process(&system_file_content, None)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn process(
-        mut self,
-        content: &[u8],
-        translation: Option<&str>,
-    ) -> Result<Option<ProcessedData>, Error> {
-        self.base.initialize_translation(translation)?;
-
-        self.system_value = parse_rpgm_file(
-            content,
-            self.base.engine_type,
-            self.base.file_type,
-        )?;
-        let mut processed = false;
-
-        for (entry_id, entry_name) in [
-            "Armor Types",
-            "Elements",
-            "Skill Types",
-            "Weapon Types",
-            "Equip Types",
-            "Terms",
-            "Currency Unit",
-            "Game Title",
-        ]
-        .into_iter()
-        .enumerate()
-        {
-            let id = entry_id as u16 + 1;
-
-            if self.base.get_translation_map(id).is_break() {
-                if self.base.mode.is_purge() {
-                    processed = true;
-                }
-
-                continue;
-            }
-
-            processed = true;
-
-            self.base.update_metadata(
-                id,
-                Vec::from([(CommentPos::Name, entry_name)]),
-            );
-
-            if id <= 5 {
-                let label = [
-                    self.base.labels.armor_types,
-                    self.base.labels.elements,
-                    self.base.labels.skill_types,
-                    self.base.labels.weapon_types,
-                    self.base.labels.equip_types,
-                ][id as usize - 1];
-
-                let Some(array) = self.system_value[label].as_array_mut()
-                else {
-                    continue;
-                };
-
-                for value in array {
-                    Self::process_value(self.base, value);
-                }
-            } else if id == 6 {
-                self.process_terms();
-            } else if id == 7 {
-                self.process_currency_unit();
-            } else {
-                self.process_game_title();
-            }
-
-            self.base.flush_translation(id);
-        }
-
-        if !processed {
-            return Ok(None);
-        }
-
-        Ok(Some(self.base.finish(self.system_value.take())))
-    }
-
-    fn process_terms(&mut self) {
-        let base = &mut *self.base;
-
-        let Some(terms) =
-            self.system_value[base.labels.terms].as_object_mut()
-        else {
-            return;
-        };
-
-        for (key, value) in terms.iter_mut() {
-            if key == "messages" {
-                if let Some(messages) = value.as_object_mut() {
-                    for value in messages.values_mut() {
-                        Self::process_value(base, value);
-                    }
-                }
-            } else if let Some(array) = value.as_array_mut() {
-                for value in array {
-                    Self::process_value(base, value);
-                }
-            } else if value.is_bytes() || value.is_string() {
-                Self::process_value(base, value);
-            }
-        }
-    }
-
-    /// Takes `base` rather than `&mut self`, so callers can hold a mutable borrow
-    /// of the disjoint `system_value` field while calling it.
-    fn process_value(base: &mut Base, value: &mut Value) {
-        let Some(extracted) = base.extract_string(&*value, true) else {
-            return;
-        };
-
-        if base.mode.is_read() {
-            base.insert_string(Cow::Borrowed(extracted));
-        } else if base.mode.is_write() {
-            if let Some(translated) = base.get_key(extracted) {
-                *value = Base::make_string_value(
-                    translated,
-                    base.engine_type.is_new(),
-                );
-            }
-        } else {
-            base.translation_map_mut()
-                .insert(extracted.into(), TranslationEntry::default());
-        }
-    }
-
-    fn process_currency_unit(&mut self) {
-        if !self.base.engine_type.is_new() {
-            let label = self.base.labels.currency_unit;
-            Self::process_value(self.base, &mut self.system_value[label]);
-        }
-    }
-
-    fn process_game_title(&mut self) {
-        if self.base.mode.is_write() {
-            if !self.game_title.is_empty() {
-                self.system_value[self.base.labels.game_title] =
-                    Value::string(self.game_title.as_str());
-            }
-        } else {
-            // User previously set the game title through set_game_title
-            if !self.game_title.is_empty() {
-                let title = take(&mut self.game_title);
-                self.base.insert_string(Cow::Owned(title));
-                return;
-            }
-
-            if let Some(game_title_value) =
-                self.system_value.get(self.base.labels.game_title)
-            {
-                let Some(game_title) =
-                    self.base.extract_string(game_title_value, true)
-                else {
-                    return;
-                };
-
-                let game_title = game_title.to_owned();
-                self.base.insert_string(Cow::Owned(game_title));
-            }
-        }
-    }
-}
-
-pub struct ScriptBase<'a> {
-    pub base: &'a mut Base,
-}
-
-impl<'a> ScriptBase<'a> {
-    /// Initializes system base using [`Base`].
-    /// Before calling this, you should create a base and pass it here.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use rvpacker_txt_rs_lib::{core::{Base, ScriptBase}, Mode, ReadMode, EngineType};
-    ///
-    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    /// let mut script_base = ScriptBase::new(&mut base);
-    /// ```
-    pub fn new(base: &'a mut Base) -> Self {
-        base.reset();
-        base.file_type = RPGMFileType::Scripts;
-
-        Self { base }
-    }
-
-    /// Processes the RPG Maker scripts file content.
-    ///
-    /// # Parameters
-    ///
-    /// - `content` - Content of the file that's being processed.
-    /// - `translation` - Contents of the translation file corresponding to the file. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if `mode` is [`Mode::Write`] and no translation exists.
-    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
-    /// - [`Error`], if unable to parse the content.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::MarshalLoad`] - if unable to load the Marshal data.
-    /// - [`Error::JsonParse`] - if unable to parse the JSON data.
-    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
-    ///
-    /// # Panics
-    ///
-    /// May panic if passed content is not `Scripts`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, ScriptBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::VXAce);
-    ///     let mut script_base = ScriptBase::new(&mut base);
-    ///
-    ///     let script_file_content = read("C:/Game/Data/Scripts.rvdata2")?;
-    ///     script_base.process(&script_file_content, None)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn process(
-        self,
-        content: &[u8],
-        translation: Option<&str>,
-    ) -> Result<Option<ProcessedData>, Error> {
-        self.base.initialize_translation(translation)?;
-
-        // SAFETY: Scripts are always array.
-        let mut scripts_array = unsafe {
-            parse_rpgm_file(
-                content,
-                self.base.engine_type,
-                self.base.file_type,
-            )?
-            .into_array()
-            .unwrap_unchecked()
-        };
-        let mut scripts = Self::decode_scripts(&scripts_array);
-
-        // SAFETY: These regexes are valid, 100% no shit.
-        let regexes = unsafe {
-            [
-                Regex::new(r"(Graphics|Data|Audio|Movies|System)\/.*\/?").unwrap_unchecked(),
-                Regex::new(r"r[xv]data2?$").unwrap_unchecked(),
-                Regex::new(r".*\(").unwrap_unchecked(),
-                Regex::new(r"^([d\d\p{P}+-]*|[d\p{P}+-]&*)$").unwrap_unchecked(),
-                Regex::new(r"^(Actor<id>|ExtraDropItem|EquipLearnSkill|GameOver|Iconset|Window|true|false|MActor%d|[wr]b|\\f|\\n|\[[A-Z]*\])$")
-                    .unwrap_unchecked(),
-            ]
-        };
-
-        let mut processed = false;
-
-        for (((script_id, script), script_name), mut code) in scripts_array
-            .iter_mut()
-            .enumerate()
-            .zip(take(&mut scripts.names))
-            .zip(take(&mut scripts.contents))
-        {
-            let id = script_id as u16 + 1;
-
-            if self.base.get_translation_map(id).is_break() {
-                if self.base.mode.is_purge() {
-                    processed = true;
-                }
-
-                continue;
-            }
-
-            processed = true;
-
-            self.base.update_metadata(
-                id,
-                Vec::from([(CommentPos::Name, script_name.as_str())]),
-            );
-            let (extracted_strings, ranges) = self.extract_strings(&code);
-
-            if self.base.mode.is_write() {
-                let mut code_changed = false;
-
-                for (extracted, range) in extracted_strings
-                    .into_iter()
-                    .zip(ranges)
-                    .filter(|(s, _)| !s.trim().is_empty())
-                    .rev()
-                {
-                    if let Some(translated) = self.base.get_key(&extracted) {
-                        code.replace_range(range, translated);
-                        code_changed = true;
-                    }
-                }
-
-                if code_changed {
-                    let mut buf = Vec::with_capacity(code.len());
-
-                    ZlibEncoder::new(&mut buf, Compression::default())
-                        .write_all(code.as_bytes())
-                        .unwrap();
-
-                    script[2] = Value::bytes(&buf);
-                }
-            } else {
-                for extracted in extracted_strings
-                    .into_iter()
-                    .filter(|s| !s.trim().is_empty())
-                {
-                    if Base::string_is_only_symbols(&extracted)
-                        || extracted.contains("@window")
-                        || extracted.contains(r"\$game")
-                        || extracted.starts_with(r"\\e")
-                        || extracted.contains("ALPHAC")
-                        || extracted.contains('_')
-                        || regexes.iter().any(|re| re.is_match(&extracted))
-                    {
-                        continue;
-                    }
-
-                    self.base.insert_string(Cow::Owned(extracted));
-                }
-
-                self.base.flush_translation(id);
-            }
-        }
-
-        if !processed {
-            return Ok(None);
-        }
-
-        Ok(Some(self.base.finish(Value::array(scripts_array))))
-    }
-
-    fn is_escaped(index: usize, string: &str) -> bool {
-        let mut backslash_count: u8 = 0;
-
-        for char in string[..index].chars().rev() {
-            if char != '\\' {
-                break;
-            }
-
-            backslash_count += 1;
-        }
-
-        backslash_count % 2 == 1
-    }
-
-    fn extract_strings(&self, ruby_code: &str) -> (Lines, Vec<Range<usize>>) {
-        let mut strings = Lines::default();
-        let mut ranges = Vec::new();
-        let mut inside_string = false;
-        let mut inside_multiline_comment = false;
-        let mut string_start_index = 0;
-        let mut current_quote_type = '\0';
-        let mut global_index = 0;
-
-        for line in ruby_code.split_inclusive('\n') {
-            let trimmed = line.trim();
-
-            if !inside_string {
-                if trimmed.starts_with('#') {
-                    global_index += line.len();
-                    continue;
-                }
-
-                if trimmed.starts_with("=begin") {
-                    inside_multiline_comment = true;
-                } else if trimmed.starts_with("=end") {
-                    inside_multiline_comment = false;
-                }
-            }
-
-            if inside_multiline_comment {
-                global_index += line.len();
-                continue;
-            }
-
-            let char_indices = line.char_indices();
-
-            for (i, char) in char_indices {
-                if !inside_string && char == '#' {
-                    break;
-                }
-
-                if !inside_string && (char == '"' || char == '\'') {
-                    inside_string = true;
-                    string_start_index = global_index + i;
-                    current_quote_type = char;
-                } else if inside_string
-                    && char == current_quote_type
-                    && !Self::is_escaped(i, &line)
-                {
-                    let range = string_start_index + 1..global_index + i;
-
-                    let extracted_string = ruby_code[range.clone()].normalize();
-
-                    if !extracted_string.is_empty()
-                        && !strings.contains(extracted_string.as_ref())
-                    {
-                        strings.insert(extracted_string.into_owned());
-
-                        if self.base.mode.is_write() {
-                            ranges.push(range);
-                        }
-                    }
-
-                    inside_string = false;
-                    current_quote_type = '\0';
-                }
-            }
-
-            global_index += line.len();
-        }
-
-        (strings, ranges)
-    }
-
-    /// Decodes an array of script entries into [`Scripts`] struct that holds `numbers`, `scripts` and `names` fields.
-    ///
-    /// # Parameters
-    ///
-    /// - `scripts_array`: Slice of script entries.
-    ///
-    /// # Returns
-    ///
-    /// A [`Scripts`] struct that holds `numbers`, `scripts` and `names` fields.
-    ///
-    /// # Panics
-    ///
-    /// May panic if decoder gets interrupted.
-    ///
-    #[must_use]
-    pub fn decode_scripts(scripts_array: &[Value]) -> Scripts {
-        let mut numbers = Vec::with_capacity(scripts_array.len());
-        let mut contents = Vec::with_capacity(scripts_array.len());
-        let mut names = Vec::with_capacity(scripts_array.len());
-
-        for script in scripts_array {
-            // SAFETY: Scripts always have a layout like this. `0` is magic number, `1` is name and `2` is actual script data.
-            let script_number = if script[0].is_bytes() {
-                unsafe {
-                    str::from_utf8_unchecked(
-                        script[0].as_byte_vec().unwrap_unchecked(),
-                    )
-                    .parse::<i32>()
-                    .unwrap_unchecked()
-                }
-            } else if script[0].is_string() {
-                unsafe {
-                    script[0]
-                        .as_str()
-                        .unwrap_unchecked()
-                        .parse::<i32>()
-                        .unwrap_unchecked()
-                }
-            } else {
-                unsafe { script[0].as_int().unwrap_unchecked() }
-            };
-            let script_name_data =
-                unsafe { script[1].as_byte_vec().unwrap_unchecked() };
-            let script_data =
-                unsafe { script[2].as_byte_vec().unwrap_unchecked() };
-
-            let mut decoded_script = Vec::with_capacity(script_data.len());
-            ZlibDecoder::new(script_data)
-                .read_to_end(&mut decoded_script)
-                .unwrap();
-
-            for encoding in [
-                encoding_rs::UTF_8,
-                encoding_rs::WINDOWS_1252,
-                encoding_rs::WINDOWS_1251,
-                encoding_rs::SHIFT_JIS,
-                encoding_rs::GB18030,
-            ] {
-                let (content_cow, _, had_errors) =
-                    encoding.decode(&decoded_script);
-                let (name_cow, _, _) = encoding.decode(script_name_data);
-
-                if !had_errors {
-                    numbers.push(script_number);
-                    contents.push(content_cow.into());
-                    names.push(name_cow.into());
-                    break;
-                }
-            }
-        }
-
-        Scripts::new(numbers, contents, names)
-    }
-
-    /// Encodes decoded [`Scripts`] struct back to [`Vec<Value>`].
-    ///
-    /// # Parameters
-    ///
-    /// - [`Scripts`] struct to encode.
-    ///
-    /// # Returns
-    ///
-    /// - [`Vec<Value>`] of encoded script entries.
-    ///
-    /// # Panics
-    ///
-    /// May panic if encoder gets interrupted.
-    ///
-    #[must_use]
-    pub fn encode_scripts(scripts: &Scripts) -> Vec<Value> {
-        let mut scripts_array = Vec::with_capacity(scripts.contents.len());
-
-        for ((content, name), number) in scripts
-            .contents
-            .iter()
-            .zip(scripts.names.iter())
-            .zip(scripts.numbers.iter())
-        {
-            let mut encoder =
-                ZlibEncoder::new(Vec::new(), Compression::default());
-            encoder.write_all(content.as_bytes()).unwrap();
-            let compressed_content = encoder.finish().unwrap();
-
-            scripts_array.push(Value::array(vec![
-                Value::int(*number),
-                Value::string(name),
-                Value::bytes(&compressed_content),
-            ]));
-        }
-
-        scripts_array
-    }
-}
-
-pub struct PluginBase<'a> {
-    pub base: &'a mut Base,
-}
-
-impl<'a> PluginBase<'a> {
-    /// Initializes system base using [`Base`].
-    /// Before calling this, you should create a base and pass it here.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use rvpacker_txt_rs_lib::{core::{Base, PluginBase}, Mode, ReadMode, EngineType};
-    ///
-    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::New);
-    /// let mut plugin_base = PluginBase::new(&mut base);
-    /// ```
-    pub fn new(base: &'a mut Base) -> Self {
-        base.reset();
-        base.file_type = RPGMFileType::Plugins;
-
-        Self { base }
-    }
-
-    /// Processes the RPG Maker plugins file content.
-    ///
-    /// # Parameters
-    ///
-    /// - `content` - Content of the file that's being processed.
-    /// - `translation` - Contents of the translation file corresponding to the file. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
-    ///
-    /// # Returns
-    ///
-    /// - Nothing if `mode` is [`Mode::Write`] and no translation exists.
-    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
-    /// - [`Error`], if unable to parse the content.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::JsonParse`] - if parsing plugin JSON content fails.
-    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
-    ///
-    /// # Panics
-    ///
-    /// May panic if passed content is not `plugins.js`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rvpacker_txt_rs_lib::{core::{Base, PluginBase}, Mode, ReadMode, EngineType, Error};
-    /// use std::fs::read;
-    ///
-    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::New);
-    ///     let mut plugin_base = PluginBase::new(&mut base);
-    ///
-    ///     let plugins_file_content = read("plugins.js")?;
-    ///     plugin_base.process(&plugins_file_content, None)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn process(
-        mut self,
-        content: &[u8],
-        translation: Option<&str>,
-    ) -> Result<Option<ProcessedData>, Error> {
-        self.base.initialize_translation(translation)?;
-
-        // SAFETY: Plugins content should always be like `plugins = [...]`, and JSON is always valid UTF-8.
-        let plugins_array_str = unsafe {
-            std::str::from_utf8_unchecked(content)
-                .split_once('=')
-                .unwrap_unchecked()
-                .1
-                .trim_end_matches([';', '\r', '\n'])
-        };
-
-        // SAFETY: Plugins are always array.
-        let mut plugins_array = unsafe {
-            Value::from(from_str::<SerdeValue>(plugins_array_str)?)
-                .into_array()
-                .unwrap_unchecked()
-        };
-
-        let mut processed = false;
-
-        for (plugin_id, plugin_object) in plugins_array.iter_mut().enumerate() {
-            let id = plugin_id as u16 + 1;
-
-            if self.base.get_translation_map(id).is_break() {
-                if self.base.mode.is_purge() {
-                    processed = true;
-                }
-
-                continue;
-            }
-
-            processed = true;
-            // SAFETY: Each plugin always contains name.
-            let plugin_name =
-                unsafe { plugin_object["name"].as_str().unwrap_unchecked() };
-
-            self.base.update_metadata(
-                id,
-                Vec::from([(CommentPos::Name, plugin_name)]),
-            );
-            self.parse_plugin(None, plugin_object);
-            self.base.flush_translation(id);
-        }
-
-        if !processed {
-            return Ok(None);
-        }
-
-        Ok(Some(self.base.finish(Value::array(plugins_array))))
-    }
-
-    fn parse_plugin(&mut self, key: Option<&str>, value: &mut Value) {
-        let is_invalid_key = |key: Option<&str>| {
-            let Some(key_string) = key else {
-                return false;
-            };
-
-            if key_string.starts_with("LATIN") {
-                false
-            } else {
-                PLUGINS_REGEXPS
-                    .with(|r| r.iter().any(|re| re.is_match(key_string)))
-            }
-        };
-
-        match &mut **value {
-            ValueType::String(value_string) => {
-                if is_invalid_key(key) {
-                    return;
-                }
-
-                if !(value_string.trim().is_empty()
-                    || IS_ONLY_SYMBOLS_RE.with(|r| r.is_match(value_string))
-                    || ["true", "false", "none", "time", "off"]
-                        .contains(&value_string.as_str())
-                    || value_string.starts_with("this.")
-                        && value_string
-                            .chars()
-                            .nth(5)
-                            .is_some_and(char::is_alphabetic)
-                        && value_string.ends_with(')')
-                    || value_string.starts_with("rgba"))
-                    || key.is_some_and(|x| x.starts_with("LATIN"))
-                {
-                    let string = value_string.normalize();
-
-                    if self.base.mode.is_write() {
-                        if let Some(translated) = self.base.get_key(&string) {
-                            *value = Value::string(translated.as_str());
-                        }
-                    } else {
-                        self.base.insert_string(string);
-                    }
-                }
-            }
-            ValueType::Object(obj) => {
-                for (key, value) in obj.iter_mut() {
-                    self.parse_plugin(Some(key), value);
-                }
-            }
-            ValueType::Array(arr) => {
-                for value in arr {
-                    self.parse_plugin(None, value);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{CustomReplace, NEW_LINE};
-    use regex::Regex;
-    use std::borrow::Cow;
-
-    /// Every case is checked against the regexes these replaced, so the
-    /// hand-rolled scans stay bug-for-bug compatible - including CRLF becoming
-    /// two separators, which falls out of the leftmost-first alternation.
-    const CASES: &[&str] = &[
-        "",
-        "plain",
-        "no breaks at all",
-        "a\nb",
-        "a\r\nb",
-        "a\rb",
-        "\n",
-        "\r\n",
-        "\r\n\r\n",
-        "a\n\nb",
-        "trailing\n",
-        "\nleading",
-        "mix\ra\nb\r\nc",
-        "\u{65E5}\u{672C}\u{8A9E}\n\u{8A9E}",
-        "\r\n\u{65E5}\u{672C}\r",
-        r"already\#done",
-        r"\#\#",
-        "a\\#b\nc",
-        "\\",
-        r"\\#",
-        "#",
-        r"\#",
-    ];
-
-    #[test]
-    fn normalize_matches_regex() {
-        let re = Regex::new(r"\r|\n|\r\n").unwrap();
-
-        for case in CASES {
-            let expected = re.replace_all(case, NEW_LINE);
-            let actual = case.normalize();
-
-            assert_eq!(actual, expected, "normalize({case:?})");
-            assert_eq!(
-                matches!(actual, Cow::Borrowed(_)),
-                matches!(expected, Cow::Borrowed(_)),
-                "normalize({case:?}) borrowed-ness"
-            );
-        }
-    }
-
-    #[test]
-    fn denormalize_matches_regex() {
-        let re = Regex::new(r"\\#").unwrap();
-
-        for case in CASES {
-            let expected = re.replace_all(case, "\n");
-            let actual = case.denormalize();
-
-            assert_eq!(actual, expected, "denormalize({case:?})");
-            assert_eq!(
-                matches!(actual, Cow::Borrowed(_)),
-                matches!(expected, Cow::Borrowed(_)),
-                "denormalize({case:?}) borrowed-ness"
-            );
-        }
-    }
-}
