+++ src/core/plugin.rs
@@ -1,0 +1,205 @@
+use super::*;
+use crate::{
+    CommentPos, ProcessedData,
+    types::{Error, RPGMFileType},
+};
+use marshal_rs::{Value, ValueType};
+use regex::Regex;
+use serde_json::{Value as SerdeValue, from_str};
+use std::cell::LazyCell;
+
+thread_local! {
+    static PLUGINS_REGEXPS: LazyCell<[Regex; 11]> = LazyCell::new(|| unsafe {
+        [
+            Regex::new(r"^(name|description|Window Width|Window Height|ATTENTION!!!|Shown Elements|Width|Outline Color|Command Alignment|Command Position|Command Rows|Chinese Font|Korean Font|Default Font|Text Align|Scenes To Draw|displacementImage|Turn Alignment|Buff Formula|Counter Alignment|Default Width|Face Indent|Fast Forward Key|Font Name|Font Name CH|Font Name KR|Name Box Padding|Name Box Added Text|Critical Rate Formula|Critical Multplier Formula|Flat Critical Formula|Default SE|---List---|Button Events List|Kill Switch|Ex Turn Image|Ex Turn Name Color|Non Ex Turn Name Color|Option menu entry|Add to options|Default Ambient Light|Reset Lights|Gab Font Name|Escape Ratio|Translated Format|Default Sound|Action Speed|Default System|Untranslated Format|Default Format|Victory Screen Level Sound|Warning Side Battle UI|Weapon Swap Text Hit|Weapon Swap Text Critical|Weapon Swap Command|Weapon Swap Text Evasion|alwaysDash|renderingMode|Attributes Command|Attributes Column 1|Attributes Column 2|Attributes Column 3|Warning OTB|</span> Minimum Damage</span></td>|Present Settings)$").unwrap_unchecked(),
+            Regex::new(r"^Folder.*\w$").unwrap_unchecked(),
+            Regex::new(r"[XY]$").unwrap_unchecked(),
+            Regex::new(r"BGM").unwrap_unchecked(),
+            Regex::new(r"Label").unwrap_unchecked(),
+            Regex::new(r"^Custom \w").unwrap_unchecked(),
+            Regex::new(r"^outlineColor").unwrap_unchecked(),
+            Regex::new(r"^(Menu|Item|Skill|Equip|Status|Save|Options|End).*(Background|Motion)$").unwrap_unchecked(),
+            Regex::new(r"^Menu \w").unwrap_unchecked(),
+            Regex::new(r"^(MHP|MMP|ATK|DEF|MAT|MDF|AGI|LUK).*(Formula|Maximum|Minimum|Effect|Color)$").unwrap_unchecked(),
+            Regex::new(r"^Damage\w*$").unwrap_unchecked(),
+        ]
+    });
+    static IS_ONLY_SYMBOLS_RE: LazyCell<Regex> = LazyCell::new(|| unsafe {
+        Regex::new(r#"^[,.()+\-:;\[\]^~%&!№$@`*\/→×？?ｘ％▼|♥♪！：〜『』「」〽。…‥＝゠、，【】［］｛｝（）〔〕｟｠〘〙〈〉《》・\\#<>=_ー※▶ⅠⅰⅡⅱⅢⅲⅣⅳⅤⅴⅥⅵⅦⅶⅧⅷⅨⅸⅩⅹⅪⅺⅫⅻⅬⅼⅭⅽⅮⅾⅯⅿ\s\d"']+$"#).unwrap_unchecked()
+    });
+}
+
+pub struct PluginBase<'a> {
+    pub base: &'a mut Base,
+}
+
+impl<'a> PluginBase<'a> {
+    /// Initializes system base using [`Base`].
+    /// Before calling this, you should create a base and pass it here.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use rvpacker_txt_rs_lib::{core::{Base, PluginBase}, Mode, ReadMode, EngineType};
+    ///
+    /// let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::New);
+    /// let mut plugin_base = PluginBase::new(&mut base);
+    /// ```
+    pub fn new(base: &'a mut Base) -> Self {
+        base.reset();
+        base.file_type = RPGMFileType::Plugins;
+
+        Self { base }
+    }
+
+    /// Processes the RPG Maker plugins file content.
+    ///
+    /// # Parameters
+    ///
+    /// - `content` - Content of the file that's being processed.
+    /// - `translation` - Contents of the translation file corresponding to the file. Isn't used with [`ReadMode::Default`]. Requires to be set with any other [`Mode`].
+    ///
+    /// # Returns
+    ///
+    /// - Nothing if `mode` is [`Mode::Write`] and no translation exists.
+    /// - [`ProcessedData`], which contains RPG Maker data if `mode` is [`Mode::Write`] and translation data otherwise.
+    /// - [`Error`], if unable to parse the content.
+    ///
+    /// # Errors
+    ///
+    /// - [`Error::JsonParse`] - if parsing plugin JSON content fails.
+    /// - [`Error::NoTranslation`] - if mode is not [`ReadMode::Default`], and no translation was passed.
+    ///
+    /// # Panics
+    ///
+    /// May panic if passed content is not `plugins.js`.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// use rvpacker_txt_rs_lib::{core::{Base, PluginBase}, Mode, ReadMode, EngineType, Error};
+    /// use std::fs::read;
+    ///
+    /// fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ///     let mut base = Base::new(Mode::Read(ReadMode::Default { force: false }), EngineType::New);
+    ///     let mut plugin_base = PluginBase::new(&mut base);
+    ///
+    ///     let plugins_file_content = read("plugins.js")?;
+    ///     plugin_base.process(&plugins_file_content, None)?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub fn process(
+        mut self,
+        content: &[u8],
+        translation: Option<&str>,
+    ) -> Result<Option<ProcessedData>, Error> {
+        self.base.initialize_translation(translation)?;
+
+        // SAFETY: Plugins content should always be like `plugins = [...]`, and JSON is always valid UTF-8.
+        let plugins_array_str = unsafe {
+            std::str::from_utf8_unchecked(content)
+                .split_once('=')
+                .unwrap_unchecked()
+                .1
+                .trim_end_matches([';', '\r', '\n'])
+        };
+
+        // SAFETY: Plugins are always array.
+        let mut plugins_array = unsafe {
+            Value::from(from_str::<SerdeValue>(plugins_array_str)?)
+                .into_array()
+                .unwrap_unchecked()
+        };
+
+        let mut processed = false;
+
+        for (plugin_id, plugin_object) in plugins_array.iter_mut().enumerate() {
+            let id = plugin_id as u16 + 1;
+
+            if self.base.get_translation_map(id).is_break() {
+                if self.base.mode.is_purge() {
+                    processed = true;
+                }
+
+                continue;
+            }
+
+            processed = true;
+            // SAFETY: Each plugin always contains name.
+            let plugin_name =
+                unsafe { plugin_object["name"].as_str().unwrap_unchecked() };
+
+            self.base.update_metadata(
+                id,
+                Vec::from([(CommentPos::Name, plugin_name)]),
+            );
+            self.parse_plugin(None, plugin_object);
+            self.base.flush_translation(id);
+        }
+
+        if !processed {
+            return Ok(None);
+        }
+
+        Ok(Some(self.base.finish(Value::array(plugins_array))))
+    }
+
+    fn parse_plugin(&mut self, key: Option<&str>, value: &mut Value) {
+        let is_invalid_key = |key: Option<&str>| {
+            let Some(key_string) = key else {
+                return false;
+            };
+
+            if key_string.starts_with("LATIN") {
+                false
+            } else {
+                PLUGINS_REGEXPS
+                    .with(|r| r.iter().any(|re| re.is_match(key_string)))
+            }
+        };
+
+        match &mut **value {
+            ValueType::String(value_string) => {
+                if is_invalid_key(key) {
+                    return;
+                }
+
+                if !(value_string.trim().is_empty()
+                    || IS_ONLY_SYMBOLS_RE.with(|r| r.is_match(value_string))
+                    || ["true", "false", "none", "time", "off"]
+                        .contains(&value_string.as_str())
+                    || value_string.starts_with("this.")
+                        && value_string
+                            .chars()
+                            .nth(5)
+                            .is_some_and(char::is_alphabetic)
+                        && value_string.ends_with(')')
+                    || value_string.starts_with("rgba"))
+                    || key.is_some_and(|x| x.starts_with("LATIN"))
+                {
+                    let string = value_string.normalize();
+
+                    if self.base.mode.is_write() {
+                        if let Some(translated) = self.base.get_key(&string) {
+                            *value = Value::string(translated.as_str());
+                        }
+                    } else {
+                        self.base.insert_string(string);
+                    }
+                }
+            }
+            ValueType::Object(obj) => {
+                for (key, value) in obj.iter_mut() {
+                    self.parse_plugin(Some(key), value);
+                }
+            }
+            ValueType::Array(arr) => {
+                for value in arr {
+                    self.parse_plugin(None, value);
+                }
+            }
+            _ => {}
+        }
+    }
+}
