+++ src/lib.rs
@@ -8,10 +8,10 @@
 
 pub mod constants;
 pub mod core;
-pub mod serde;
 pub mod generic;
 pub mod json;
 pub mod processors;
+pub mod serde;
 pub mod types;
 
 pub use constants::{
